@@ -1,310 +1,289 @@
 [build-system]
-requires = ["setuptools>=61.2", "setuptools_scm[toml]>=7.0"]
-build-backend = "setuptools.build_meta"
+build-backend = "hatchling.build"
+requires = [
+    "hatch-fancy-pypi-readme",
+    "hatch-vcs",
+    "hatchling",
+]
 
 [project]
 name = "tmmc-lnpy"
-authors = [{ name = "William P. Krekelberg", email = "wpk@nist.gov" }]
-license = { text = "NIST-PD" }
 description = "Analysis of lnPi results from TMMC simulation"
-# if using markdown
-# long_description_content_type = text/markdown
-keywords = ["tmmc-lnpy"]
+keywords = [
+    "tmmc-lnpy",
+]
+license = { text = "NIST-PD" }
+license-files = { paths = ["LICENSE"] }
+authors = [{ name = "William P. Krekelberg", email = "wpk@nist.gov" }]
+requires-python = ">=3.8,<=3.12"
 classifiers = [
     "Development Status :: 2 - Pre-Alpha",
+    "Intended Audience :: Science/Research",
     "License :: Public Domain",
     "Operating System :: OS Independent",
-    "Intended Audience :: Science/Research",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering",
 ]
-dynamic = ["readme", "version"]
-requires-python = ">=3.8"
+dynamic = [
+    "readme",
+    "version",
+]
 dependencies = [
-    "numpy",
-    "xarray",
-    "scipy",
-    "scikit-image",
-    "joblib",
     "bottleneck",
-    "tqdm",
     "importlib_resources",
-    "module-utilities >= 0.9.0",
+    "joblib",
     "lazy_loader",
-] # additional packages
-
-[project.urls]
-Homepage = "https://github.com/usnistgov/tmmc-lnpy"
-Documentation = "https://pages.nist.gov/tmmc-lnpy/"
-
+    "module-utilities >= 0.9.0",
+    "numpy",
+    "scikit-image >= 0.21",
+    "scipy",
+    "tqdm",
+    "typing-extensions;python_version<'3.11'",
+    "xarray",
+]
 [project.optional-dependencies]
-test = [
-    "pytest", #
-    "pytest-xdist",
-    "pytest-cov",
-    "pytest-sugar",
-]
-viz = [
-    "matplotlib-base", #
-    "ipywidgets",
-]
-dev-extras = [
-    "setuptools-scm", #
-    "pytest-accept", # p2c: -p
-    "ipython",
-    "ipykernel",
-    "tmmc-lnpy[viz]",
-    "nbval",
-]
-typing-extras = [
-    "pytype; python_version < '3.11'", #
-    "mypy >= 1.4.1",
-    "pandas-stubs",
-    "types-tqdm",
-]
-typing = [
-    "tmmc-lnpy[typing-extras]", #
-    "tmmc-lnpy[test]",
-]
-nox = [
-    "nox",
-    "noxopt", # p2c: -p
-    "ruamel.yaml",
-]
 dev = [
-    "tmmc-lnpy[test]", #
-    "tmmc-lnpy[typing-extras]",
-    "tmmc-lnpy[dev-extras]",
-]
-tools = [
-    "pre-commit", #
-    "cruft",
-    "scriv",
-    "nbqa",
+    "tmmc-lnpy[dev-extras,test,typing,pipxrun,mypy,viz,nbval]",
 ]
 dev-complete = [
-    "tmmc-lnpy[dev]", #
-    "tmmc-lnpy[tools]",
-    "tmmc-lnpy[nox]",
+    "tmmc-lnpy[dev,tools]",
+]
+dev-extras = [
+    "ipykernel",
+    "ipython",
+    "nbclient >= 0.10.0",
+    "nox >= 2024.3.2",
+    "pytest-accept",
 ]
 docs = [
-    "ipython", #
+    "autodocsumm",
+    "myst-nb",
     "pyenchant",
-    "ghp-import",
     "sphinx >= 5.3.0",
+    "sphinx-autobuild",
+    "sphinx-book-theme",
     "sphinx-copybutton",
     "sphinxcontrib-spelling",
-    "sphinx-autobuild",
-    "myst-nb",
-    "sphinx-book-theme",
-    "autodocsumm",
     "tmmc-lnpy[viz]",
 ]
-# to be parsed with pyproject2conda with --no-base option
-dist-pypi = ["twine", "build"]
-dist-conda = [
-    "anaconda-client", #
-    "grayskull",
-    "conda-build",
-    "conda-verify",
-    "boa",
+mypy = [
+    "mypy >= 1.9.0",
+]
+nbval = [
+    "nbval",
+]
+pipxrun = [
+    # Anything needed to run tools/pipxrun.py
+    "packaging",
+]
+pipxrun-tools = [
+    # these used solely for specking tools to be used with tools/pipxrun.py
+    "conda-lock >= 2.5.5",
+    "grayskull >= 2.5.3",
+    "nbqa >= 1.8.4",
+    "pyright >= 1.1.356",
+    "tmmc-lnpy[mypy]",
+    "twine >= 5.0.0",
+]
+test = [
+    "pytest",
+    "pytest-cov",
+    "pytest-sugar",
+    "pytest-xdist",
+]
+tools = [
+    "pipx",
+    "pre-commit",
+    "scriv",
+]
+typing = [
+    "pandas-stubs",
+    "pytest",
+    "tmmc-lnpy[pipxrun]",
+    "types-tqdm",
+]
+viz = [
+    "ipywidgets",
+    "matplotlib",
+]
+[project.urls]
+Documentation = "https://pages.nist.gov/tmmc-lnpy/"
+Homepage = "https://github.com/usnistgov/tmmc-lnpy"
+
+[tool.hatch.version]
+source = "vcs"
+
+[tool.hatch.build]
+exclude = [
+    ".*_cached",
 ]
 
-[tool.pyproject2conda]
-channels = ["conda-forge"]
+[tool.hatch.build.targets.sdist]
+exclude = [
+    "examples",
+    "docs",
+]
 
-## grayskull still messes some things up, but use scripts/recipe-append.sh for this
-[tool.setuptools]
-zip-safe = true # if using mypy, must be False
-include-package-data = true
-license-files = ["LICENSE"]
+[tool.hatch.build.targets.wheel]
+packages = ["src/lnpy", "src/lnPi"]
 
-[tool.setuptools.packages.find]
-namespaces = true
-where = ["src"]
 
-## include = []
-## exclude = []
-[tool.setuptools.dynamic]
-readme = { file = [
-    "README.md",
-    "CHANGELOG.md",
-    "LICENSE"
-], content-type = "text/markdown" }
+[tool.hatch.metadata.hooks.fancy-pypi-readme]
+content-type = "text/markdown"
+fragments = [
+    { path = "README.md" },
+    { path = "CHANGELOG.md" },
+    { path = "LICENSE" },
+]
 
-[tool.setuptools_scm]
-fallback_version = "999"
-write_to = "src/lnpy/_version.py"
-
-[tool.aliases]
-test = "pytest"
-
-[tool.pytest.ini_options]
-addopts = "--doctest-modules --doctest-glob='*.md' --pyargs"
-testpaths = ["tests", "lnpy", "lnPi", "README.md"]
-
-[tool.coverage.report]
-exclude_also = ["if TYPE_CHECKING:", "@overload"]
-omit = ["*/vendored/*"]
-include = ["src/*"]
-
-[tool.isort]
-profile = "black"
-skip_gitignore = true
-known_first_party = ["lnpy"]
+# * Linting --------------------------------------------------------------------
 
 [tool.ruff]
 fix = true
 line-length = 88
-target-version = "py38"
-select = [
-    # pyflakes
-    "F",
-    # pycodestyle
-    "E",
-    "W",
-    # isort
-    "I",
-    # pyupgrade
-    "UP",
-    # pydocstyle
-    "D",
-    # # flake8-2020
-    "YTT",
-    # # flake8-bugbear
-    # "B",
-    # flake8-quotes
-    "Q",
-    # # pylint
-    # "PLE", "PLR", "PLW",
-    # # misc lints
-    "PIE",
-    # # tidy imports
-    "TID",
-    # # implicit string concatenation
-    # "ISC",
-    # # type-checking imports
-    "TCH",
-]
-# Allow autofix for all enabled rules (when `--fix`) is provided.
-# fixable = ["A", "B", "C", "D", "E", "F", "..."]
-unfixable = []
-# Exclude a variety of commonly ignored directories.
-exclude = [
-    ".bzr",
-    ".direnv",
-    ".eggs",
-    ".git",
-    ".hg",
-    ".mypy_cache",
-    ".nox",
-    ".pants.d",
-    ".pytype",
-    ".ruff_cache",
-    ".svn",
-    ".tox",
+extend-include = ["*.ipynb"]
+extend-exclude = [
     ".venv",
-    "__pypackages__",
-    "_build",
-    "buck-out",
-    "build",
-    "dist",
-    "node_modules",
-    "venv",
-    "tests/",
-    "src/lnpy/tests",
+    "dist-conda",
+    ".numba_cache",
+    "examples/archived",
+    # "tests/",
+    # "src/lnpy/tests",
 ]
+unsafe-fixes = true
+
+[tool.ruff.lint]
+preview = true
+select = ["ALL"]
+# fixable = ["ALL"]
+unfixable = ["T"]
 ignore = [
-    # # whitespace before ':' - doesn't work well with black
-    # "E203",
-    # module level import not at top of file
-    "E402",
-    # line too long - let black worry about that
-    "E501",
-    # do not assign a lambda expression, use a def
-    "E731",
-    # # line break before binary operator
-    # "W503",
-    # allow black line after docstring
-    "D202",
-    "D105",
-    "D205",
-    # this leads to errors with placing titles in module
-    # docstrings
-    "D400",
-    "D401",
-    "D415",
-    "D102",
-    "D103",
-    # these are useful, but too many errors
-    # due to use of docfiller
-    "D417",
-    "D107",
-    "D203",
-    "D212",
-    # Allow relative imports
-    "TID252",
+    "CPY",       # - Don't require copyright
+    "ERA",       # - eradicate (want to keep some stuff)
+    "FBT",       # - bools are ok
+    # * Annotations
+    "ANN101",    # - Leads to issues with methods and self
+    "ANN102",    # - Leads to issues with classmethods and cls
+    "ANN401",    # - Any ok sometimes
+    # * Allow non top level imports
+    "PLC0415",   # - import should be at top level (leads to issues with imports in func?)
+    "TID252",    # - Allow relative imports
+    "E402",      # - module level import not at top of file
+    # * Other
+    "E501",      # - line too long - let formatter fix this
+    "E731",      # - do not assign a lambda expression, use a def
+    # * Docstrings
+    "D105",      # - Missing magic method docstring
+    "D205",      # - blank line after summary
+    # * This leads to errors with placing titles in module
+    "D400",      # - First line should end with a period
+    "D401",      # - First line of docstring should be in imperative mood: "{first_line}"
+    "D415",      # - First line should end with a period, question mark, or exclamation point
+    # * Not a fan of these.  And lead to issues with docfiller
+    "D107",      # - Missing docstring in __init__
+    "D203",      # - 1 blank line required before class docstring
+    "D212",      # - Multi-line docstring summary should start at the first line
+    "D417",      # - Missing argument description in the docstring for {definition}: {name} (bad with docfiller)
+    # * These conflict with formatter
+    "COM",       # - flake8-commas, formatter should take care of this?
+    "ISC",       # - flake8-implicit-str-concat
 ]
-per-file-ignores = {  }
 # Allow unused variables when underscore-prefixed.
 dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
-
-[tool.ruff.isort]
-known-first-party = ["lnpy"]
-
-[tool.ruff.mccabe]
-# Unlike Flake8, default to a complexity level of 10.
-max-complexity = 10
-
-[tool.nbqa.addopts]
-ruff = ["--fix", "--extend-ignore=D100,B018"]
-
-[tool.flake8]
-docstring-convention = "numpy"
-ignore = [
-    # # whitespace before ':' - doesn't work well with black
-    "E203",
-    # module level import not at top of file
-    "E402",
-    # line too long - let black worry about that
-    "E501",
-    # do not assign a lambda expression, use a def
-    "E731",
-    # # line break before binary operator
-    "W503",
-    # allow black line after docstring
-    "D202",
-    "D105",
-    "D205",
-    # this leads to errors with placing titles in module
-    # docstrings
-    "D400",
-    "D401",
-    "D415",
-    "D102",
-    "D103",
-    # these are useful, but too many errors
-    # due to use of docfiller
-    "D417",
-    "D107",
-    "D203",
-    "D212",
+# Use extend-ignore for project specific ignores (avoid changing ignore directly)
+extend-ignore = [
+    # "PD",        # - not using pandas?
+    # "NPY",       # - not using numpy?
+    # * Annotations
+    "SLF001",  # - flake8-self (private access sometimes OK)
+    # "ANN",     # - Annotations (just use mypy/pyright)
+    # * pylint
+    # "PLR2004", # - numbers in comparison sometimes ok
+    # "PLR0913", # - complexity sometimes ok
+    # "PLR0917", # - complexity sometimes ok
+    # * Docstrings ymmv
+    "D102",    # - Missing docstring in public method
+    "D103",    # - Missing docstring in public function
+    "D202",    # - blank line after docstring
+    # * ignore complexity for now
+    "C901",
+    "PLR091",
 ]
 
-[tool.scriv]
-format = "md"
-md_header_level = "2"
-new_fragment_template = "file: changelog.d/templates/new_fragment.md.j2"
+[tool.ruff.lint.pydocstyle]
+convention = "numpy"
 
-[tool.commitizen]
-use_shortcuts = true
+[tool.ruff.lint.pylint]
+max-args = 15
+max-positional-args = 15
+max-public-methods = 30
 
-[tool.cruft]
+[tool.ruff.lint.per-file-ignores]
+"docs/conf.py" = ["INP001"]
+"noxfile.py" = ["RUF009"]
+"tools/*.py" = ["S", "A", "SLF001", "D102"]
+"tests/*.py" = ["D", "S101"]
+"**/*.ipynb" = ["D100", "B018", "INP001"]
+
+[tool.ruff.lint.extend-per-file-ignores]
+# Project specific ignores (avoid changing per-file-ignores directly)
+"tests/*.py" = ["ANN"]
+"**/conftest.py" = ["PT", "ANN"]
+"**/*.py" = ["N999", "ANN101", "ANN102", "ANN401", "PLR2004", "PLC0415", "TD003", "FIX002"]
+"**/lnPi/*.py" = ["PGH", "ANN", "PL", "B", "N", "SIM", "S101"]
+"**/ensembles.py" = ["N802", "N803", "N806"]
+"**/*.ipynb" = ["ANN", "T201", "PLR2004", "S110", "PD010", "PD013", "E302", "E305"]
+
+[tool.ruff.lint.isort]
+known-first-party = ["lnpy", "lnPi"]
+
+[tool.ruff.format]
+docstring-code-format = true
+
+[tool.nbqa.exclude]
+codespell = "^examples/archived"
+
+[tool.check-wheel-contents]
+toplevel = ["lnpy", "lnPi"]
+src_dir = ["src"]
+
+# * Testing --------------------------------------------------------------------
+
+[tool.pytest.ini_options]
+addopts = [
+    "--doctest-modules",
+    "--doctest-glob=*.md",
+    "--cov-config=pyproject.toml",
+    "--pyargs",
+    "-n",
+    "4",
+]
+testpaths = [
+    # "README.md",  # py38 issues.
+    "lnPi",
+    "lnpy",
+    "tests",
+]
+
+[tool.coverage.run]
+branch = true
+
+[tool.coverage.report]
+# show_missing = true
+exclude_also = ["if TYPE_CHECKING:", "@overload"]
+omit = ["*/vendored/*", "*/_typing.py"]
+include = ["src/*"]
+
+[tool.coverage.paths]
+source = ["src/", ".nox/**/site-packages/"]
+
+# * Typing ---------------------------------------------------------------------
 
 [tool.mypy]
 files = ["src", "tests"]
@@ -362,3 +346,69 @@ reportUnboundVariable = false
 
 [tool.pytype]
 inputs = ["src", "tests"]
+
+# * Dependency creation --------------------------------------------------------
+
+[tool.pyproject2conda]
+user_config = "config/userconfig.toml"
+template_python = "requirements/py{py}-{env}"
+template = "requirements/{env}"
+python = ["3.11"]
+channels = ["conda-forge"]
+style = ["yaml", "requirements"]
+default_envs = ["test", "typing", "docs", "dev", "dev-complete"]
+
+[tool.pyproject2conda.dependencies]
+pytest-accept = { pip = true }
+scriv = { pip = true }
+matplotlib = { skip = true, packages = "matplotlib-base" }
+
+[tool.pyproject2conda.envs.build]
+extras = "build-system.requires"
+reqs = ["build"]
+base = false
+style = "requirements"
+
+[tool.pyproject2conda.envs.pipxrun-tools]
+extras = "pipxrun-tools"
+base = false
+style = "requirements"
+
+[tool.pyproject2conda.envs.test-extras]
+extras = ["test"]
+base = false
+
+[tool.pyproject2conda.envs.test-notebook]
+extras = ["test", "nbval", "viz"]
+
+[[tool.pyproject2conda.overrides]]
+# Ensure these have pip in yaml file
+envs = ["test", "docs", "dev", "dev-complete"]
+deps = ["pip"]
+
+[[tool.pyproject2conda.overrides]]
+envs = ["test", "typing", "test-extras"]
+python = ["3.8", "3.9", "3.10", "3.11", "3.12"]
+
+[tool.pip-tools]
+allow-unsafe = true
+strip-extras = true
+
+# * Other tools ----------------------------------------------------------------
+
+[tool.scriv]
+format = "md"
+md_header_level = "2"
+new_fragment_template = "file: changelog.d/templates/new_fragment.md.j2"
+
+[tool.commitizen]
+use_shortcuts = true
+
+[tool.cruft]
+
+[tool.typos.default.extend-words]
+datas = "datas"
+
+[tool.typos.default.extend-identifiers]
+MyNDArray = "MyNDArray"
+fileprefix_inh = "fileprefix_inh"
