@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 #
 # python_boilerplate documentation build configuration file, created by
 # sphinx-quickstart on Fri Jun  9 13:47:02 2017.
@@ -18,10 +17,15 @@
 # absolute, like shown here.
 #
 """Build docs."""
+
+from __future__ import annotations
+
 import os
 import sys
+from pathlib import Path
+from typing import Any
 
-sys.path.insert(0, os.path.abspath(".."))
+sys.path.insert(0, str(Path("../src").resolve()))
 
 import lnpy
 
@@ -41,23 +45,24 @@ extensions = [
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.autosectionlabel",
-    "IPython.sphinxext.ipython_directive",
-    "IPython.sphinxext.ipython_console_highlighting",
+    # "IPython.sphinxext.ipython_directive",
+    # "IPython.sphinxext.ipython_console_highlighting",
     # "nbsphinx",
-    ## easier external links
+    # - easier external links
     # "sphinx.ext.extlinks",
-    ## view source code on created page
+    # - view source code on created page
     # "sphinx.ext.viewcode",
-    ## view source code on github
+    # - view source code on github
     "sphinx.ext.linkcode",
-    ## add copy button
+    # - add copy button
     "sphinx_copybutton",
-    ## redirect stuff?
+    # - redirect stuff?
     # "sphinxext.rediraffe",
-    ## pretty things up?
+    # - pretty things up?
     # "sphinx_design"
-    ## myst stuff
+    # - myst stuff
     "myst_nb",
+    # "myst_parser",
 ]
 
 nitpicky = True
@@ -88,7 +93,7 @@ myst_enable_extensions = [
 ]
 
 
-myst_heading_anchors = 2
+myst_heading_anchors = 3
 myst_footnote_transition = True
 myst_dmath_double_inline = True
 myst_enable_checkboxes = True
@@ -121,13 +126,16 @@ nb_kernel_rgx_aliases = {"lnpy.*": "python3", "conda.*": "python3"}
 
 nb_execution_allow_errors = True
 
+# Whether to remove stderr
+nb_output_stderr = "remove"
+
 # - top level variables --------------------------------------------------------
 # set github_username variable to be subbed later.
 # this makes it easy to switch from wpk -> usnistgov later
 github_username = "usnistgov"
 
 html_context = {
-    "github_user": "usnistgov",
+    "github_user": github_username,
     "github_repo": "tmmc-lnpy",
     "github_version": "main",
     "doc_path": "docs",
@@ -248,32 +256,17 @@ master_doc = "index"
 
 # General information about the project.
 project = "tmmc-lnpy"
-# copyright = "2022, William P. Krekelberg"
+copyright = "2015, William P. Krekelberg"  # noqa: A001
 author = "William P. Krekelberg"
+
 
 # The version info for the project you're documenting, acts as replacement
 # for |version| and |release|, also used in various other places throughout
 # the built documents.
 #
 # The short X.Y version.
-# versioning with scm with editable install has issues.
-# instead, try to use scm if available.
-# try:
-#     from setuptools_scm import get_version
-
-#     version = get_version(root="..", relative_to=__file__)
-#     release = version
-# except ImportError:
-#     version = lnpy.__version__
-#     # The full version, including alpha/beta/rc tags.
-#     release = lnpy.__version__
-
-
-def _get_version():
-    import os
-
-    version = os.environ.get("SETUPTOOLS_SCM_PRETEND_VERSION", None)
-    if version is None:
+def _get_version() -> str:
+    if (version := os.environ.get("SETUPTOOLS_SCM_PRETEND_VERSION")) is None:
         version = lnpy.__version__
     return version
 
@@ -312,19 +305,19 @@ todo_include_todos = False
 
 html_theme = "sphinx_book_theme"
 
-html_theme_options = dict(
-    # analytics_id=''  this is configured in rtfd.io
-    # canonical_url="",
-    repository_url=f"https://github.com/{github_username}/tmmc-lnpy",
-    repository_branch=html_context["github_version"],
-    path_to_docs=html_context["doc_path"],
-    # use_edit_page_button=True,
-    use_repository_button=True,
-    use_issues_button=True,
-    home_page_in_toc=True,
-    show_toc_level=1,
-    show_navbar_depth=1,
-)
+html_theme_options = {
+    # "analytics_id": ''  this is configured in rtfd.io
+    # "canonical_url": "",
+    "repository_url": f"https://github.com/{github_username}/tmmc-lnpy",
+    "repository_branch": html_context["github_version"],
+    "path_to_docs": html_context["doc_path"],
+    # "use_edit_page_button": True,
+    "use_repository_button": True,
+    "use_issues_button": True,
+    "home_page_in_toc": True,
+    "show_toc_level": 1,
+    "show_navbar_depth": 1,
+}
 # handle nist css/js from here.
 html_css_files = [
     # "css/nist-combined.css",
@@ -354,12 +347,14 @@ html_static_path = ["_static"]
 # Sometimes the savefig directory doesn't exist and needs to be created
 # https://github.com/ipython/ipython/issues/8733
 # becomes obsolete when we can pin ipython>=5.2; see ci/requirements/doc.yml
-ipython_savefig_dir = os.path.join(
-    os.path.dirname(os.path.abspath(__file__)), "_build", "html", "_static"
-)
-if not os.path.exists(ipython_savefig_dir):
-    os.makedirs(ipython_savefig_dir)
+def _get_ipython_savefig_dir() -> str:
+    d = Path(__file__).parent / "_build" / "html" / "_static"
+    if not d.is_dir():
+        d.mkdir(parents=True)
+    return str(d)
 
+
+ipython_savefig_dir = _get_ipython_savefig_dir()
 
 # If not '', a 'Last updated on:' timestamp is inserted at every page bottom,
 # using the given strftime format.
@@ -463,8 +458,8 @@ linkcheck_ignore = [
 
 
 # based on numpy doc/source/conf.py
-def linkcode_resolve(domain, info):
-    """Determine the URL corresponding to Python object"""
+def linkcode_resolve(domain: str, info: dict[str, Any]) -> str | None:
+    """Determine the URL corresponding to Python object."""
     import inspect
     from operator import attrgetter
 
@@ -499,12 +494,11 @@ def linkcode_resolve(domain, info):
     except OSError:
         lineno = None
 
-    if lineno:
-        linespec = f"#L{lineno}-L{lineno + len(source) - 1}"
-    else:
-        linespec = ""
+    linespec = f"#L{lineno}-L{lineno + len(source) - 1}" if lineno else ""
 
-    fn = os.path.relpath(fn, start=os.path.dirname(lnpy.__file__))
+    # fmt: off
+    fn = os.path.relpath(fn, start=Path(lnpy.__file__).parent)
+    # fmt: on
 
     return f"https://github.com/{github_username}/tmmc-lnpy/blob/{html_context['github_version']}/src/lnpy/{fn}{linespec}"
 
