@@ -1,662 +1,849 @@
 """Config file for nox."""
+
+# * Imports ----------------------------------------------------------------------------
 from __future__ import annotations
 
+import shlex
 import shutil
-from dataclasses import replace  # noqa
-from itertools import product
+import sys
+from functools import lru_cache, partial, wraps
+
+# Should only use on python version > 3.10
+if sys.version_info < (3, 10):
+    msg = "python>=3.10 required"
+    raise RuntimeError(msg)
+
+from dataclasses import dataclass
 from pathlib import Path
-from textwrap import dedent
 from typing import (
+    TYPE_CHECKING,
     Annotated,
     Any,
     Callable,
-    Collection,
+    Iterable,
     Iterator,
     Literal,
     Sequence,
     TypeAlias,
-    TypeVar,
-    cast,
+    TypedDict,
 )
 
-import nox
-from noxopt import NoxOpt, Option, Session
-
-from tools.noxtools import (  # session_install_package,
+# fmt: off
+sys.path.insert(0, ".")
+from tools import pipxrun
+from tools.dataclass_parser import DataclassParser, add_option, option
+from tools.noxtools import (
+    Installer,
+    cached_which,
+    check_for_change_manager,
+    combine_list_list_str,
     combine_list_str,
+    factory_conda_backend,
+    factory_virtualenv_backend,
+    infer_requirement_path,
+    is_conda_session,
     load_nox_config,
     open_webpage,
     prepend_flag,
-    session_install_envs,
-    session_install_envs_lock,
-    session_install_pip,
     session_run_commands,
-    sort_like,
     update_target,
 )
 
-# * nox options ------------------------------------------------------------------------
-ROOT = Path(__file__).parent
+sys.path.pop(0)
 
-nox.options.reuse_existing_virtualenvs = True
-nox.options.sessions = ["test"]
+# make sure these afeter
+import nox  # type: ignore[unused-ignore,import]
 
-# * Options ----------------------------------------------------------------------------
+# fmt: on
+
+if TYPE_CHECKING:
+    from nox import Session
+    from nox.virtualenv import CondaEnv
+
+
+# * Names ------------------------------------------------------------------------------
 
 PACKAGE_NAME = "tmmc-lnpy"
 IMPORT_NAME = "lnpy"
 KERNEL_BASE = "lnpy"
 
-PYTHON_ALL_VERSIONS = ["3.8", "3.9", "3.10", "3.11"]
-PYTHON_DEFAULT_VERSION = "3.10"
+# * nox options ------------------------------------------------------------------------
 
-# conda/mamba
-if shutil.which("mamba"):
-    CONDA_BACKEND = "mamba"
-elif shutil.which("conda"):
-    CONDA_BACKEND = "conda"  # pyright: ignore
-else:
-    raise ValueError("neither conda or mamba found")
+ROOT = Path(__file__).parent
 
-SESSION_DEFAULT_KWS = {"python": PYTHON_DEFAULT_VERSION, "venv_backend": CONDA_BACKEND}
-SESSION_ALL_KWS = {"python": PYTHON_ALL_VERSIONS, "venv_backend": CONDA_BACKEND}
+nox.options.reuse_existing_virtualenvs = True
+nox.options.sessions = ["test"]
+nox.options.default_venv_backend = "uv"
 
-
-# * User config ------------------------------------------------------------------------
+# * User Config ------------------------------------------------------------------------
 
 CONFIG = load_nox_config()
+# if you'd like to disallow uv.
+# You'll need to import this from tools.noxtools
+# DISALLOW_WHICH.append("uv")
 
-# * noxopt -----------------------------------------------------------------------------
-group = NoxOpt(auto_tag=True)
 
-F = TypeVar("F", bound=Callable[..., Any])
-C: TypeAlias = Callable[[F], F]
+# * Options ---------------------------------------------------------------------------
 
-DEFAULT_SESSION = cast(C[F], group.session(**SESSION_DEFAULT_KWS))  # type: ignore
-ALL_SESSION = cast(C[F], group.session(**SESSION_ALL_KWS))  # type: ignore
+LOCK = True
 
-DEFAULT_SESSION_VENV = cast(C[F], group.session(python=PYTHON_DEFAULT_VERSION))  # type: ignore
-ALL_SESSION_VENV = cast(C[F], group.session(python=PYTHON_ALL_VERSIONS))  # type: ignore
+PYTHON_ALL_VERSIONS = ["3.8", "3.9", "3.10", "3.11", "3.12"]
+PYTHON_DEFAULT_VERSION = "3.11"
 
-OPTS_OPT = Option(nargs="*", type=str)
-# SET_KERNEL_OPT = Option(type=bool, help="If True, try to set the kernel name")
-RUN_OPT = Option(
-    nargs="*",
-    type=str,
-    action="append",
-    help="run passed command_demo using `external=True` flag",
+
+PIPXRUN_LOCK_REQUIREMENTS = "requirements/lock/py{}-pipxrun-tools.txt".format(
+    PYTHON_DEFAULT_VERSION.replace(".", "")
 )
-
-CMD_OPT = Option(nargs="*", type=str, help="cmd to be run")
-LOCK_OPT = Option(type=bool, help="If True, use conda-lock")
+PIPXRUN_MIN_REQUIREMENTS = "requirements/pipxrun-tools.txt"
 
 
-def opts_annotated(**kwargs: Any):  # type: ignore
-    return Annotated[list[str], replace(OPTS_OPT, **kwargs)]
+@lru_cache
+def get_pipxrun_specs(requirements: str | None = None) -> pipxrun.Specifications:
+    """Get specs for pipxrun."""
+    requirements = requirements or PIPXRUN_MIN_REQUIREMENTS
+    return pipxrun.Specifications.from_requirements(requirements=requirements)
 
 
-def cmd_annotated(**kwargs):  # type: ignore
-    return Annotated[list[str], replace(CMD_OPT, **kwargs)]
+for backend in ["mamba", "micromamba", "conda"]:
+    if shutil.which(backend):
+        CONDA_BACKEND: Literal["mamba", "micromamba", "conda"] = backend  # type: ignore[assignment]
+        break
+else:
+    msg = "no conda-like backend found"
+    raise ValueError(msg)
 
 
-def run_annotated(**kwargs):  # type: ignore
-    return Annotated[list[list[str]], replace(RUN_OPT, **kwargs)]
+class SessionOptionsDict(TypedDict, total=False):
+    """Dict for options to nox.session"""
+
+    python: str | list[str]
+    venv_backend: str | Callable[..., CondaEnv]
 
 
-LOCK_CLI = Annotated[bool, LOCK_OPT]
-RUN_CLI = Annotated[list[list[str]], RUN_OPT]
-TEST_OPTS_CLI = opts_annotated(help="extra arguments/flags to pytest")
+CONDA_DEFAULT_KWS: SessionOptionsDict = {
+    "python": PYTHON_DEFAULT_VERSION,
+    "venv_backend": factory_conda_backend(CONDA_BACKEND),
+}
+CONDA_ALL_KWS: SessionOptionsDict = {
+    "python": PYTHON_ALL_VERSIONS,
+    "venv_backend": factory_conda_backend(CONDA_BACKEND),
+}
 
-# CMD_CLI = Annotated[list[str], CMD_OPT]
+DEFAULT_KWS: SessionOptionsDict = {"python": PYTHON_DEFAULT_VERSION}
+ALL_KWS: SessionOptionsDict = {"python": PYTHON_ALL_VERSIONS}
 
-FORCE_REINSTALL_CLI = Annotated[
-    bool,
-    Option(
-        type=bool,
-        help="If True, force reinstall requirements and package even if environment unchanged",
+
+# * Session command line options -------------------------------------------------------
+
+OPT_TYPE: TypeAlias = list[str] | None
+RUN_TYPE: TypeAlias = list[list[str]] | None
+
+RUN_ANNO = Annotated[
+    RUN_TYPE,
+    option(
+        help="Run external commands in session.  Specify multiple times for multiple commands.",
     ),
 ]
-
-VERSION_CLI = Annotated[
-    str, Option(type=str, help="Version to substitute or check against")
-]
-
-LOG_SESSION_CLI = Annotated[
-    bool,
-    Option(
-        type=bool,
-        help="If flag included, log python and package (if installed) version",
-    ),
-]
+OPT_ANNO = Annotated[OPT_TYPE, option(help="Options to command.")]
 
 
-# * Installation command ---------------------------------------------------------------
-def py_prefix(python_version: Any) -> str:
-    if isinstance(python_version, str):
-        return "py" + python_version.replace(".", "")
-    else:
-        raise ValueError(f"passed non-string value {python_version}")
+@dataclass
+class SessionParams(DataclassParser):
+    """Holds all cli options"""
+
+    # common parameters
+    lock: bool = False
+    update: bool = add_option("--update", "-U", help="update dependencies/package")
+    update_package: bool = add_option(
+        "--update-package",
+        "-P",
+        help="update package only",
+    )
+    log_session: bool = add_option("--log-session")
+    version: str | None = None
+
+    # dev
+    dev_run: RUN_ANNO = None
+    dev_envname: Literal["dev", "dev-complete", "dev-user"] = add_option(
+        help="Name of environment to use for development session",
+        default="dev",
+    )
+
+    # config
+    dev_extras: OPT_TYPE = add_option(help="`extras` to include in dev environment")
+    python_paths: OPT_TYPE = add_option(help="paths to python executables")
+
+    # requirements
+    requirements_force: bool = False
+    requirements_no_notify: bool = add_option(
+        default=False,
+        help="Skip notification of pip-compile",
+    )
+
+    # conda-lock
+    conda_lock_channel: OPT_TYPE = add_option(help="conda channels")
+    conda_lock_platform: (
+        list[Literal["osx-64", "linux-64", "win-64", "osx-arm64", "all"]] | None
+    ) = add_option(help="platform(s) to buiuld lock file for.")
+    conda_lock_include: OPT_TYPE = add_option(help="lock files to create")
+    conda_lock_mamba: bool = False
+    conda_lock_force: bool = False
+
+    # pip-compile
+    pip_compile_force: bool = False
+    pip_compile_upgrade: bool = add_option(
+        "--pip-compile-upgrade",
+        "-L",
+        help="Upgrade all packages in lock file",
+        default=False,
+    )
+    pip_compile_upgrade_package: OPT_TYPE = add_option(
+        help="Upgrade package(s) in lock file",
+        default=None,
+    )
+    pip_compile_opts: OPT_TYPE = add_option(help="options to pip-compile")
+    pip_compile_run: RUN_ANNO = None
+    pip_compile_run_internal: RUN_ANNO = None
+
+    # test
+    test_no_pytest: bool = False
+    test_opts: OPT_TYPE = add_option(help="Options to pytest")
+    test_run: RUN_ANNO = None
+    no_cov: bool = False
+
+    # coverage
+    coverage: list[Literal["erase", "combine", "report", "html", "open"]] | None = None
+
+    # testdist
+    testdist_run: RUN_ANNO = None
+
+    # docs
+    docs: (
+        list[
+            Literal[
+                "html",
+                "build",
+                "symlink",
+                "clean",
+                "livehtml",
+                "linkcheck",
+                "spelling",
+                "showlinks",
+                "open",
+                "serve",
+            ]
+        ]
+        | None
+    ) = add_option("--docs", "-d", help="doc commands")
+    docs_run: RUN_ANNO = None
+
+    # typing
+    typing: list[
+        Literal[
+            "clean",
+            "mypy",
+            "pyright",
+            "pytype",
+            "all",
+            "mypy-notebook",
+            "pyright-notebook",
+            "typecheck-notebook",
+        ]
+    ] = add_option("--typing", "-m")
+    typing_run: RUN_ANNO = None
+    typing_run_internal: RUN_TYPE = add_option(
+        help="Run internal (in session) commands.",
+    )
+
+    # build
+    build: list[Literal["build", "version"]] | None = None
+    build_run: RUN_ANNO = None
+    build_isolation: bool = False
+    build_outdir: str = "./dist"
+    build_opts: OPT_ANNO = None
+    build_silent: bool = False
+
+    # publish
+    publish: list[Literal["release", "test", "check"]] | None = add_option(
+        "-p", "--publish"
+    )
+
+    # conda-recipe/grayskull
+    conda_recipe: list[Literal["recipe", "recipe-full"]] | None = None
+    conda_recipe_sdist_path: str | None = None
+
+    # conda-build
+    conda_build: list[Literal["build", "clean"]] | None = None
+    conda_build_run: RUN_ANNO = None
 
 
-def session_environment_filename(
-    name: str | None,
-    ext: str | None = None,
-    python_version: str | None = None,
-) -> str:
-    if name is None:
-        raise ValueError("must supply name")
+@lru_cache
+def parse_posargs(*posargs: str) -> SessionParams:
+    """
+    Get Parser using `+` for session option prefix.
 
-    filename = name
-    if ext is not None:
-        filename = filename + ext
-    if python_version is not None:
-        filename = f"{py_prefix(python_version)}-{filename}"
-    return f"./environment/{filename}"
+    Note that using `+` allows for passing underlying `-` options
+    without escaping.
+    """
+    opts = SessionParams.from_posargs(posargs=posargs, prefix_char="+")
+    opts.lock = opts.lock or LOCK
+
+    return opts
 
 
-def pkg_install_condaenv(
-    session: nox.Session,
-    name: str,
-    lock: bool = False,
-    display_name: str | None = None,
-    install_package: bool = True,
-    force_reinstall: bool = False,
-    log_session: bool = False,
-    deps: Collection[str] | None = None,
-    reqs: Collection[str] | None = None,
-    channels: Collection[str] | None = None,
-    filename: str | None = None,
-    **kwargs: Any,
-) -> None:
-    """Install requirements.  If need fine control, do it in calling func."""
+def add_opts(
+    func: Callable[[Session, SessionParams], None],
+) -> Callable[[Session], None]:
+    """Fill in `opts` from cli options."""
 
-    def check_filename(filename: str | Path) -> str:
-        if not Path(filename).exists():
-            raise ValueError(f"file {filename} does not exist")
-        session.log(f"Environment file: {filename}")
-        return str(filename)
+    @wraps(func)
+    def wrapped(session: Session) -> None:
+        opts = parse_posargs(*session.posargs)
+        return func(session, opts)
 
-    if lock:
-        filename = (
-            filename
-            or f"./environment/lock/{py_prefix(session.python)}-{name}-conda-lock.yml"
-        )
-        session_install_envs_lock(
-            session=session,
-            lockfile=check_filename(filename),
-            display_name=display_name,
-            force_reinstall=force_reinstall,
-            install_package=install_package,
-            **kwargs,
-        )
-
-    else:
-        assert isinstance(session.python, str)
-
-        filename = filename or session_environment_filename(
-            name=name,
-            ext=".yaml",
-            python_version=session.python,
-        )
-        session_install_envs(
-            session,
-            check_filename(filename),
-            display_name=display_name,
-            force_reinstall=force_reinstall,
-            deps=deps,
-            reqs=reqs,
-            channels=channels,
-            install_package=install_package,
-            **kwargs,
-        )
-
-    if log_session:
-        session_log_session(session, install_package)
-
-
-def pkg_install_venv(
-    session: nox.Session,
-    name: str,  # pyright: ignore
-    lock: bool = False,
-    requirement_paths: Collection[str] | None = None,
-    constraint_paths: Collection[str] | None = None,
-    extras: str | Collection[str] | None = None,
-    reqs: Collection[str] | None = None,
-    display_name: str | None = None,
-    force_reinstall: bool = False,
-    install_package: bool = False,
-    no_deps: bool = False,
-    log_session: bool = False,
-) -> None:
-    if lock:
-        raise ValueError("lock not yet supported for install_pip")
-
-    else:
-        session_install_pip(
-            session=session,
-            requirement_paths=requirement_paths,
-            constraint_paths=constraint_paths,
-            extras=extras,
-            reqs=reqs,
-            display_name=display_name,
-            force_reinstall=force_reinstall,
-            install_package=install_package,
-            no_deps=no_deps,
-        )
-
-    if log_session:
-        session_log_session(session, install_package)
-
-
-def session_log_session(session: nox.Session, has_package: bool = False) -> None:
-    session.run("python", "--version")
-    if has_package:
-        session.run(
-            "python",
-            "-c",
-            dedent(
-                f"""
-        import {IMPORT_NAME}
-        print({IMPORT_NAME}.__version__)
-        """
-            ),
-        )
+    return wrapped
 
 
 # * Environments------------------------------------------------------------------------
-# ** Development (conda)
-@DEFAULT_SESSION
+# ** Dev (conda)
+@add_opts
 def dev(
     session: Session,
-    dev_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Create dev env."""
-    # using conda
-
-    pkg_install_condaenv(
-        session=session,
-        name="dev",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-dev",
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
+    """Create development environment using either conda (dev) or virtualenv (dev-venv) in location `.venv`"""
+    (
+        Installer.from_envname(
+            session=session,
+            envname=opts.dev_envname,
+            lock=opts.lock,
+            update=opts.update,
+            package=True,
+        )
+        .install_all(
+            update_package=opts.update_package,
+            log_session=opts.log_session,
+        )
+        .set_ipykernel_display_name(name=f"{PACKAGE_NAME}-dev", update=True)
+        .run_commands(opts.dev_run)
     )
-    session_run_commands(session, dev_run)
 
 
-@DEFAULT_SESSION_VENV
-def dev_venv(
+nox.session(
+    name="dev-venv",
+    python=PYTHON_DEFAULT_VERSION,
+    venv_backend=factory_virtualenv_backend(location="./.venv"),
+)(dev)
+nox.session(
+    name="dev",
+    python=PYTHON_DEFAULT_VERSION,
+    venv_backend=factory_conda_backend(backend=CONDA_BACKEND, location="./.venv"),
+)(dev)
+
+
+@nox.session(python=False)
+@add_opts
+def config(
     session: Session,
-    dev_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Create dev env."""
-    # using conda
+    """Create the file ./config/userconfig.toml"""
+    args: list[str] = []
+    if opts.dev_extras:
+        args += ["--dev-extras", *opts.dev_extras]
+    if opts.python_paths:
+        args += ["--python-paths", *opts.python_paths]
 
-    pkg_install_venv(
-        session=session,
-        name="dev-venv",
-        lock=lock,
-        extras=["dev"],
-        display_name=f"{PACKAGE_NAME}-dev-venv",
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
-    session_run_commands(session, dev_run)
+    session.run(sys.executable, "tools/projectconfig.py", *args)
 
 
-# ** pyproject2conda (create environment.yaml and requirement.txt files)
-@DEFAULT_SESSION_VENV
-def pyproject2conda(
+# ** requirements
+@nox.session(name="requirements", python=False)
+@add_opts
+def requirements(
     session: Session,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    pyproject2conda_force: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Create environment.yaml files from pyproject.toml using pyproject2conda."""
-    pkg_install_venv(
+    """
+    Create environment.yaml and requirement.txt files from pyproject.toml using pyproject2conda.
+
+    These will be placed in the directory "./requirements".
+    """
+    pipxrun.run(
+        "pyproject2conda>=0.11.0",
+        "project",
+        "--verbose",
+        *(["--overwrite=force"] if opts.requirements_force else []),
         session=session,
-        name="pyporject2conda",
-        reqs=["pyproject2conda>=0.4.0"],
-        force_reinstall=force_reinstall,
+        external=True,
     )
 
-    def create_env(
-        python_version: str | None = None,
-        cmd: Literal["yaml", "requirements"] = "yaml",
-        name: str | None = None,
-        output: str | None = None,
-        extras: str | Sequence[str] | None = None,
-        python_include: str | bool = True,
-        base: bool = True,
-    ) -> None:
-        def _to_args(flag: str, val: str | Sequence[str] | None) -> list[str]:
-            if val is None:
-                return []
-            if isinstance(val, str):
-                val = [val]
-            return prepend_flag(flag, *val)
-
-        if output is None:
-            assert name is not None
-            output = session_environment_filename(
-                python_version=python_version,
-                name=name,
-                ext={"yaml": ".yaml", "requirements": ".txt"}[cmd],
-            )
-
-        if pyproject2conda_force or update_target(output, "pyproject.toml"):
-            args = [cmd, "-o", output] + _to_args("-e", extras)
-
-            if cmd == "yaml":
-                if python_version is not None:
-                    args.extend(["--python-version", python_version])
-                if isinstance(python_include, bool) and python_include:
-                    python_include = f"python={python_version}"
-                if isinstance(python_include, str):
-                    args.extend(["--python-include", python_include])
-
-            if not base:
-                args.append("--no-base")
-
-            session.run("pyproject2conda", *args)
+    if not opts.requirements_no_notify and opts.lock:
+        if cached_which("uv"):
+            session.notify("uv-compile")
         else:
-            session.log(
-                f"{output} up to data.  Pass --pyproject2conda-force to force recreation"
-            )
-
-    extras = CONFIG["environment-extras"]
-
-    # All versions:
-    for env, python_version in product(["test", "typing"], PYTHON_ALL_VERSIONS):
-        create_env(
-            name=env,
-            extras=extras.get(env, env),
-            base=True,
-            python_version=python_version,
-        )
-
-    for env, python_version in product(["docs", "dev"], [PYTHON_DEFAULT_VERSION]):
-        create_env(
-            name=env,
-            extras=extras.get(env, env),
-            base=True,
-            python_version=python_version,
-        )
-
-    # need an isolated set of test requirements
-    for python_version in PYTHON_ALL_VERSIONS:
-        create_env(
-            name="test-extras",
-            extras="test",
-            base=False,
-            python_version=python_version,
-        )
-
-    # isolated
-    for env in ["dist-pypi", "dist-conda"]:
-        create_env(
-            name=f"{env}",
-            extras=env,
-            base=False,
-            python_version=PYTHON_DEFAULT_VERSION,
-        )
-
-    # isolated requirement files.
-    # no python versioning for these
-    for env, extras in [("test-extras", "test"), ("dist-pypi", "dist-pypi")]:
-        create_env(name=f"{env}", cmd="requirements", extras=extras, base=False)
+            for py in PYTHON_ALL_VERSIONS:
+                session.notify(f"pip-compile-{py}")
 
 
-# ** conda-lock
-@DEFAULT_SESSION_VENV
+# # ** conda-lock
+@nox.session(name="conda-lock", python=False)
+@add_opts
 def conda_lock(
     session: Session,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    conda_lock_channel: cmd_annotated(help="conda channels to use") = (),  # type: ignore
-    conda_lock_platform: cmd_annotated(  # type: ignore
-        help="platforms to build lock files for",
-        choices=["osx-64", "linux-64", "win-64", "all"],
-    ) = (),
-    conda_lock_cmd: cmd_annotated(  # type: ignore
-        help="lock files to create",
-        choices=["test", "typing", "dev", "dist-pypi", "dist-conda", "all"],
-    ) = (),
-    conda_lock_run: RUN_CLI = [],  # noqa
-    conda_lock_mamba: bool = False,
-    conda_lock_force: bool = False,
+    opts: SessionParams,
 ) -> None:
     """Create lock files using conda-lock."""
+    conda_lock_exclude = ["test-extras"]
+    conda_lock_include = opts.conda_lock_include or [
+        "test",
+        "dev",
+        "dev-complete",
+        "nox",
+    ]
 
-    pkg_install_venv(
-        session,
-        name="conda-lock",
-        reqs=["conda-lock>=2.0.0"],
-        force_reinstall=force_reinstall,
-    )
+    platform = opts.conda_lock_platform
+    if platform is None or "all" in platform:
+        # for now, skip osx-arm64 and win-64.  Leads to some problems.
+        platform = ["osx-64", "linux-64"]  # , "win-64"]
 
-    session.run("conda-lock", "--version")
-
-    platform = cast(Sequence[str], conda_lock_platform)
-    if not platform:
-        platform = ["osx-64"]
-    elif "all" in platform:
-        platform = ["linux-64", "osx-64", "win-64"]
-    channel = cast(Sequence[str], conda_lock_channel)
+    channel = opts.conda_lock_channel
     if not channel:
         channel = ["conda-forge"]
 
-    lock_dir = ROOT / "environment" / "lock"
+    def create_lock(path: Path) -> None:
+        name = path.with_suffix("").name
+        lockfile = path.parent / "lock" / f"{name}-conda-lock.yml"
+        deps = [str(path)]
 
-    def create_lock(
-        py: str,
-        name: str,
-        env_path: str | None = None,
-    ) -> None:
-        py = "py" + py.replace(".", "")
+        # check if skip
+        env = "-".join(name.split("-")[1:])
+        if conda_lock_include and not any(c == env for c in conda_lock_include):
+            session.log(f"Skipping {lockfile} (include)")
+            return
 
-        if env_path is None:
-            env_path = f"environment/{py}-{name}.yaml"
+        if conda_lock_exclude and any(c == env for c in conda_lock_exclude):
+            session.log(f"Skipping {lockfile} (exclude)")
+            return
 
-        lockfile = lock_dir / f"{py}-{name}-conda-lock.yml"
+        # check hashes
 
-        deps = [env_path]
-        # make sure this is last to make python version last
-        # deps.append(lock_dir / f"{py}.yaml")
+        with check_for_change_manager(
+            *deps,
+            target_path=lockfile,
+            force_write=opts.conda_lock_force,
+        ) as changed:
+            if opts.conda_lock_force or changed:
+                session.log(f"Creating {lockfile}")
+                # insert -f for each arg
+                if lockfile.exists():
+                    lockfile.unlink()
+                pipxrun.run(
+                    "conda-lock",
+                    "--mamba" if opts.conda_lock_mamba else "--no-mamba",
+                    *prepend_flag("-c", *channel),
+                    *prepend_flag("-p", *platform),
+                    *prepend_flag("-f", *deps),
+                    f"--lockfile={lockfile}",
+                    specs=get_pipxrun_specs(),
+                    session=session,
+                    external=True,
+                )
+            else:
+                session.log(f"Skipping {lockfile} (exists)")
 
-        if conda_lock_force or update_target(lockfile, *deps):
-            session.log(f"creating {lockfile}")
-            # insert -f for each arg
-            if lockfile.exists():
-                lockfile.unlink()
+    for path in (ROOT / "requirements").relative_to(ROOT.cwd()).glob("py*.yaml"):
+        create_lock(path)
+
+
+def _run_compile_pre(
+    runner: Installer,
+    run: RUN_TYPE,
+    run_internal: RUN_TYPE,
+) -> bool:
+    if run:
+        runner.run_commands(run)
+        return True
+
+    if run_internal:
+        runner.run_commands(run_internal, external=False)
+        return True
+
+    return False
+
+
+def _run_compile_options(
+    options: OPT_TYPE,
+    force: bool,
+    upgrade: bool,
+    upgrade_package: OPT_TYPE,
+) -> tuple[list[str], bool]:
+    options = list(options) if options else []
+    if upgrade:
+        options = [*options, "-U"]
+
+    if upgrade_package:
+        options = [*options, *prepend_flag("-P", upgrade_package)]
+
+    force = force or upgrade or bool(upgrade_package)
+
+    return options, force
+
+
+def _run_compile_env(
+    session: nox.Session,
+    compile_command: Sequence[str],
+    python: str,
+    options: Sequence[str],
+    force: bool,
+    env: str,
+    envs_optional: Sequence[str],
+    **kwargs: Any,
+) -> None:
+    reqspath = infer_requirement_path(env, ext=".txt", check_exists=False)
+    if not reqspath.is_file():
+        if env in envs_optional:
+            return
+        msg = f"Missing file {reqspath}"
+        raise ValueError(msg)
+
+    lockpath = infer_requirement_path(
+        env,
+        ext=".txt",
+        python_version=python,
+        lock=True,
+        check_exists=False,
+    )
+
+    with check_for_change_manager(
+        reqspath,
+        target_path=lockpath,
+        force_write=force,
+    ) as changed:
+        if force or changed:
+            session.log(f"Creating {lockpath}")
             session.run(
-                "conda-lock",
-                "--mamba" if conda_lock_mamba else "--no-mamba",
-                *prepend_flag("-c", *channel),
-                *prepend_flag("-p", *platform),
-                *prepend_flag("-f", *deps),
-                f"--lockfile={lockfile}",
+                *compile_command, *options, "-o", str(lockpath), str(reqspath), **kwargs
             )
 
-    session_run_commands(session, conda_lock_run)
+        else:
+            session.log(f"Skipping {lockpath}")
 
-    if not conda_lock_run and not conda_lock_cmd:
-        conda_lock_cmd = ["all"]  # pyright: ignore
-    if "all" in conda_lock_cmd:
-        conda_lock_cmd = ["test", "typing", "dev", "dist-pypi", "dist-conda"]
-    conda_lock_cmd = list(set(conda_lock_cmd))
 
-    for c in conda_lock_cmd:
-        if c == "test":
-            for py in PYTHON_ALL_VERSIONS:
-                create_lock(py, "test")
-        elif c == "typing":
-            for py in PYTHON_ALL_VERSIONS:
-                create_lock(py, "typing")
-        elif c == "dev":
-            create_lock(PYTHON_DEFAULT_VERSION, "dev")
-        elif c == "dist-pypi":
-            create_lock(
-                PYTHON_DEFAULT_VERSION,
-                "dist-pypi",
-            )
-        elif c == "dist-conda":
-            create_lock(
-                PYTHON_DEFAULT_VERSION,
-                "dist-conda",
+@nox.session(name="pip-compile", **ALL_KWS)
+@add_opts
+def pip_compile(
+    session: Session,
+    opts: SessionParams,
+) -> None:
+    """
+    Run pip-compile.
+
+    Note that this session is also used to run pip-sync with correct python version for
+    tests/typing/etc.
+    """
+    runner = Installer(
+        session=session,
+        pip_deps=["pip-tools"],
+        update=opts.update,
+    ).install_all(log_session=opts.log_session)
+
+    if not isinstance(session.python, str):
+        msg = "must set python version"
+        raise TypeError(msg)
+
+    if _run_compile_pre(runner, opts.pip_compile_run, opts.pip_compile_run_internal):
+        return
+
+    options, force = _run_compile_options(
+        options=opts.pip_compile_opts,
+        force=opts.pip_compile_force,
+        upgrade=opts.pip_compile_upgrade,
+        upgrade_package=opts.pip_compile_upgrade_package,
+    )
+
+    envs_all = ["test", "typing"]
+    envs_dev = ["dev", "dev-complete", "docs"]
+    envs_dev_optional = ["test-notebook", "pipxrun-tools"]
+
+    if session.python == PYTHON_DEFAULT_VERSION:
+        envs = envs_all + envs_dev + envs_dev_optional
+    else:
+        envs = envs_all
+
+    for env in envs:
+        _run_compile_env(
+            session=session,
+            compile_command=("pip-compile",),
+            python=session.python,
+            options=options,
+            force=force,
+            env=env,
+            envs_optional=envs_dev_optional,
+            external=False,
+        )
+
+
+# ** uv pip compile
+@nox.session(name="uv-compile", python=False)
+@add_opts
+def uv_compile(
+    session: Session,
+    opts: SessionParams,
+) -> None:
+    """Run uv pip compile ..."""
+    uv_path = cached_which("uv")
+    if uv_path is None:
+        session.log("Need to install uv to use it...")
+        return
+
+    options, force = _run_compile_options(
+        options=opts.pip_compile_opts,
+        force=opts.pip_compile_force,
+        upgrade=opts.pip_compile_upgrade,
+        upgrade_package=opts.pip_compile_upgrade_package,
+    )
+
+    envs_all = ["test", "typing"]
+    envs_dev = ["dev", "dev-complete", "docs"]
+    envs_dev_optional = ["test-notebook", "pipxrun-tools"]
+
+    for python in set(PYTHON_ALL_VERSIONS).union({PYTHON_DEFAULT_VERSION}):
+        if python == PYTHON_DEFAULT_VERSION:
+            envs = envs_all + envs_dev + envs_dev_optional
+        else:
+            envs = envs_all
+
+        for env in envs:
+            _run_compile_env(
+                session=session,
+                compile_command=(
+                    uv_path,
+                    "pip",
+                    "compile",
+                    f"--python-version={python}",
+                    "--annotation-style=line",
+                ),
+                python=python,
+                options=options,
+                force=force,
+                env=env,
+                envs_optional=envs_dev_optional,
+                external=True,
             )
 
 
 # ** testing
 def _test(
     session: nox.Session,
-    run: list[list[str]],
+    run: RUN_TYPE,
     test_no_pytest: bool,
-    test_opts: list[str],
+    test_opts: OPT_TYPE,
     no_cov: bool,
 ) -> None:
+    import os
+
+    tmpdir = os.environ.get("TMPDIR", None)
+
     session_run_commands(session, run)
     if not test_no_pytest:
-        opts = combine_list_str(test_opts)
+        opts = combine_list_str(test_opts or [])
         if not no_cov:
             session.env["COVERAGE_FILE"] = str(Path(session.create_tmp()) / ".coverage")
-            if "--cov" not in opts:
-                opts.append("--cov")
+
+            if not any(o.startswith("--cov") for o in opts):
+                opts.append(f"--cov={IMPORT_NAME}")
+
+        # Because we are testing if temporary folders
+        # have git or not, we have to make sure we're above the
+        # not under this repo
+        # so revert to using the top level `TMPDIR`
+        if tmpdir:
+            session.env["TMPDIR"] = tmpdir
+
         session.run("pytest", *opts)
 
 
-@ALL_SESSION
+# *** Basic tests
+@add_opts
 def test(
     session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    test_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
-    no_cov: bool = False,
+    opts: SessionParams,
 ) -> None:
     """Test environments with conda installs."""
-
-    pkg_install_condaenv(
-        session=session,
-        name="test",
-        lock=lock,
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test",
+            lock=opts.lock,
+            # To use editable install
+            # package=True,
+            # To use full install
+            package=get_package_wheel(session, opts="--no-deps --force-reinstall"),
+            update=opts.update,
+        ).install_all(log_session=opts.log_session, update_package=opts.update_package)
     )
 
     _test(
         session=session,
-        run=test_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=no_cov,
+        run=opts.test_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=opts.test_opts,
+        no_cov=opts.no_cov,
     )
 
 
-@ALL_SESSION_VENV
-def test_venv(
+nox.session(name="test", **ALL_KWS)(test)
+nox.session(name="test-conda", **CONDA_ALL_KWS)(test)
+
+
+@nox.session(name="test-notebook", **DEFAULT_KWS)
+@add_opts
+def test_notebook(session: nox.Session, opts: SessionParams) -> None:
+    """Run pytest --nbval."""
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test-notebook",
+            lock=opts.lock,
+            package=get_package_wheel(session, opts="--no-deps --force-reinstall"),
+            update=opts.update,
+        ).install_all(log_session=opts.log_session, update_package=opts.update_package)
+    )
+
+    test_nbval_opts = shlex.split(
+        """
+    --nbval
+    --nbval-current-env
+    --nbval-sanitize-with=config/nbval.ini
+    --dist loadscope
+     examples/usage
+   """,
+    )
+
+    test_opts = (opts.test_opts or []) + test_nbval_opts
+
+    session.log(f"{test_opts = }")
+
+    _test(
+        session=session,
+        run=opts.test_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=test_opts,
+        no_cov=opts.no_cov,
+    )
+
+
+@nox.session(python=False)
+@add_opts
+def coverage(
     session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    test_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,  # pyright: ignore
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
-    no_cov: bool = False,
+    opts: SessionParams,
 ) -> None:
-    """Test environments virtualenv and pip installs."""
+    """Run coverage."""
+    cmd = opts.coverage or ["combine", "html", "report"]
 
-    pkg_install_venv(
-        session=session,
-        name="test-venv",
-        extras="test",
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
+    run = partial(pipxrun.run, specs=get_pipxrun_specs(), session=session)
 
-    _test(
-        session=session,
-        run=test_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=no_cov,
-    )
+    paths = list(Path(".nox").glob("test-*/tmp/.coverage*"))
 
-
-# ** coverage
-def _coverage(
-    session: nox.Session,
-    run: list[list[str]],
-    cmd: list[str],
-    run_internal: list[list[str]],
-) -> None:
-    session_run_commands(session, run)
-
-    if not cmd and not run and not run_internal:
-        cmd = ["combine", "report"]
-
-    session.log(f"{cmd}")
+    if "erase" in cmd:
+        for path in paths:
+            if path.exists():
+                session.log(f"removing {path}")
+                path.unlink()
 
     for c in cmd:
         if c == "combine":
-            paths = list(Path(".nox").glob("test-3*/tmp/.coverage"))
             if update_target(".coverage", *paths):
-                session.run("coverage", "combine", "--keep", "-a", *map(str, paths))
+                run(
+                    "coverage",
+                    "combine",
+                    "--keep",
+                    "-a",
+                    *paths,
+                )
         elif c == "open":
             open_webpage(path="htmlcov/index.html")
+
         else:
-            session.run("coverage", c)
+            run(
+                "coverage",
+                c,
+            )
 
-    session_run_commands(session, run_internal, external=False)
 
-
-@DEFAULT_SESSION_VENV
-def coverage(
+# *** testdist (conda)
+def testdist(
     session: Session,
-    coverage_cmd: cmd_annotated(  # type: ignore
-        choices=["erase", "combine", "report", "html", "open"]
-    ) = (),
-    coverage_run: RUN_CLI = [],  # noqa
-    coverage_run_internal: run_annotated(  # type: ignore
-        help="Arbitrary commands to run within the session"
-    ) = [],  # noqa
-    force_reinstall: FORCE_REINSTALL_CLI = False,
 ) -> None:
-    pkg_install_venv(
-        session,
-        name="coverage",
-        reqs=["coverage[toml]"],
-        force_reinstall=force_reinstall,
+    """Test conda distribution."""
+    opts = parse_posargs(*session.posargs)
+
+    install_str = PACKAGE_NAME
+    if opts.version:
+        install_str = f"{install_str}=={opts.version}"
+
+    if is_conda_session(session):
+        pip_deps, conda_deps = None, install_str
+    else:
+        pip_deps, conda_deps = install_str, None
+
+    (
+        Installer.from_envname(
+            session=session,
+            envname="test-extras",
+            conda_deps=conda_deps,
+            pip_deps=pip_deps,
+            update=opts.update,
+            channels="conda-forge",
+        ).install_all(log_session=opts.log_session)
     )
 
-    _coverage(
+    _test(
         session=session,
-        run=coverage_run,
-        cmd=cast(list[str], coverage_cmd),
-        run_internal=cast(list[list[str]], coverage_run_internal),
+        run=opts.testdist_run,
+        test_no_pytest=opts.test_no_pytest,
+        test_opts=opts.test_opts,
+        no_cov=opts.no_cov,
     )
 
 
-# ** Docs
-def _docs(
-    session: nox.Session, run: list[list[str]], cmd: list[str], version: str
+nox.session(name="testdist-pypi", **ALL_KWS)(testdist)
+nox.session(name="testdist-conda", **CONDA_ALL_KWS)(testdist)
+
+
+# # ** Docs
+@add_opts
+def docs(
+    session: nox.Session,
+    opts: SessionParams,
 ) -> None:
-    if version:
-        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = version
+    """
+    Run `make` in docs directory.
 
-    session_run_commands(session, run)
+    For example, 'nox -s docs -- +d html'
+    calls 'make -C docs html'. With 'release' option, you can set the
+    message with 'message=...' in posargs.
+    """
+    runner = Installer.from_envname(
+        session=session,
+        envname="docs",
+        lock=opts.lock,
+        package=True,
+        update=opts.update,
+    ).install_all(
+        update_package=opts.update_package,
+        log_session=opts.log_session,
+    )
 
-    if not run and not cmd:
+    if opts.version:
+        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = opts.version
+
+    runner.run_commands(opts.docs_run)
+
+    cmd = opts.docs or []
+    if not opts.docs_run and not cmd:
         cmd = ["html"]
 
     if "symlink" in cmd:
@@ -666,311 +853,77 @@ def _docs(
     if open_page := "open" in cmd:
         cmd.remove("open")
 
+    if serve := "serve" in cmd:
+        open_webpage(url="http://localhost:8000")
+        cmd.remove("serve")
+
     if cmd:
-        args = ["make", "-C", "docs"] + combine_list_str(cmd)
+        args = ["make", "-C", "docs", *combine_list_str(cmd)]
         session.run(*args, external=True)
 
     if open_page:
         open_webpage(path="./docs/_build/html/index.html")
 
-
-@DEFAULT_SESSION
-def docs(
-    session: nox.Session,
-    docs_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "html",
-            "build",
-            "symlink",
-            "clean",
-            "livehtml",
-            "linkcheck",
-            "spelling",
-            "showlinks",
-            "release",
-            "open",
-        ],
-        flags=("--docs-cmd", "-d"),
-    ) = (),
-    docs_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Runs make in docs directory. For example, 'nox -s docs -- --docs-cmd html' -> 'make -C docs html'. With 'release' option, you can set the message with 'message=...' in posargs."""
-    pkg_install_condaenv(
-        session=session,
-        name="docs",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-docs",
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
-
-    _docs(
-        session=session, cmd=docs_cmd, run=docs_run, version=version
-    )  # pyright: ignore
-
-
-@DEFAULT_SESSION_VENV
-def docs_venv(
-    session: nox.Session,
-    docs_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "html",
-            "build",
-            "symlink",
-            "clean",
-            "livehtml",
-            "linkcheck",
-            "spelling",
-            "showlinks",
-            "release",
-            "open",
-        ],
-        flags=("--docs-cmd", "-d"),
-    ) = (),
-    docs_run: RUN_CLI = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Runs make in docs directory. For example, 'nox -s docs -- --docs-cmd html' -> 'make -C docs html'. With 'release' option, you can set the message with 'message=...' in posargs."""
-    pkg_install_venv(
-        session=session,
-        name="docs-venv",
-        lock=lock,
-        display_name=f"{PACKAGE_NAME}-docs-venv",
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-        extras="docs",
-    )
-
-    _docs(
-        session=session, cmd=docs_cmd, run=docs_run, version=version
-    )  # pyright: ignore
-
-
-# ** Dist pypi
-def _dist_pypi(
-    session: nox.Session, run: list[list[str]], cmd: list[str], version: str
-) -> None:
-    if version:
-        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = version
-    session_run_commands(session, run)
-    if not run and not cmd:
-        cmd = ["build"]
-    if cmd:
-        if "build" in cmd:
-            cmd.append("clean")
-        cmd = sort_like(cmd, ["clean", "build", "testrelease", "release"])
-
-        session.log(f"cmd={cmd}")
-
-        for command in cmd:
-            if command == "clean":
-                session.run("rm", "-rf", "dist", external=True)
-            elif command == "build":
-                session.run("python", "-m", "build", "--outdir", "dist/")
-
-            elif command == "testrelease":
-                session.run("twine", "upload", "--repository", "testpypi", "dist/*")
-
-            elif command == "release":
-                session.run("twine", "upload", "dist/*")
-
-
-@DEFAULT_SESSION_VENV
-def dist_pypi(
-    session: nox.Session,
-    dist_pypi_run: RUN_CLI = [],  # noqa
-    dist_pypi_cmd: cmd_annotated(  # type: ignore
-        choices=["clean", "build", "testrelease", "release"],
-        flags=("--dist-pypi-cmd", "-p"),
-    ) = (),
-    lock: LOCK_CLI = False,  # pyright: ignore
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Run 'nox -s dist-pypi -- {clean, build, testrelease, release}'."""
-
-    pkg_install_venv(
-        session=session,
-        name="dist-pypi",
-        requirement_paths=[session_environment_filename(name="dist-pypi.txt")],
-        force_reinstall=force_reinstall,
-        install_package=False,
-        log_session=log_session,
-    )
-
-    _dist_pypi(
-        session=session,
-        run=dist_pypi_run,
-        cmd=dist_pypi_cmd,  # pyright: ignore
-        version=version,
-    )
-
-
-@DEFAULT_SESSION
-def dist_pypi_condaenv(
-    session: nox.Session,
-    dist_pypi_run: RUN_CLI = [],  # noqa
-    dist_pypi_cmd: cmd_annotated(  # type: ignore
-        choices=["clean", "build", "testrelease", "release"],
-        flags=("--dist-pypi-cmd", "-p"),
-    ) = (),
-    lock: LOCK_CLI = False,  # pyright: ignore
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Run 'nox -s dist_pypi -- {clean, build, testrelease, release}'."""
-    # conda
-
-    pkg_install_condaenv(
-        session=session,
-        name="dist-pypi",
-        install_package=False,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
-
-    _dist_pypi(
-        session=session,
-        run=dist_pypi_run,
-        cmd=dist_pypi_cmd,  # pyright: ignore
-        version=version,
-    )
-
-
-# ** Dist conda
-@DEFAULT_SESSION
-def dist_conda(
-    session: nox.Session,
-    dist_conda_run: RUN_CLI = [],  # noqa
-    dist_conda_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "recipe",
-            "build",
-            "clean",
-            "clean-recipe",
-            "clean-build",
-            "recipe-cat-full",
-        ],
-        flags=("--dist-conda-cmd", "-c"),
-    ) = (),
-    # lock: LOCK_CLI = False,
-    sdist_path: str = "",
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
-    version: VERSION_CLI = "",
-) -> None:
-    """Runs make -C dist-conda posargs."""
-    pkg_install_condaenv(
-        session=session,
-        name="dist-conda",
-        install_package=False,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
-
-    run, cmd = dist_conda_run, dist_conda_cmd
-    session_run_commands(session, run)
-    if not run and not cmd:
-        cmd = ["recipe"]
-
-    if cmd:
-        if "recipe" in cmd:
-            cmd.append("clean-recipe")
-        if "build" in cmd:
-            cmd.append("clean-build")
-        if "clean" in cmd:
-            cmd.extend(["clean-recipe", "clean-build"])
-            cmd.remove("clean")
-
-        cmd = sort_like(
-            cmd, ["recipe-cat-full", "clean-recipe", "recipe", "clean-build", "build"]
+    if serve and "livehtml" not in cmd:
+        session.run(
+            "python",
+            "-m",
+            "http.server",
+            "-d",
+            "docs/_build/html",
+            "-b",
+            "127.0.0.1",
+            "8000",
         )
 
-        if not sdist_path:
-            sdist_path = PACKAGE_NAME
-            if version:
-                sdist_path = f"{sdist_path}=={version}"
 
-        for command in cmd:
-            if command == "clean-recipe":
-                session.run("rm", "-rf", f"dist-conda/{PACKAGE_NAME}", external=True)
-            elif command == "clean-build":
-                session.run("rm", "-rf", "dist-conda/build", external=True)
-            elif command == "recipe":
-                session.run(
-                    "grayskull",
-                    "pypi",
-                    sdist_path,
-                    "--sections",
-                    "package",
-                    "source",
-                    "build",
-                    "requirements",
-                    "-o",
-                    "dist-conda",
-                )
-                _append_recipe(
-                    f"dist-conda/{PACKAGE_NAME}/meta.yaml", ".recipe-append.yaml"
-                )
-                session.run(
-                    "cat", f"dist-conda/{PACKAGE_NAME}/meta.yaml", external=True
-                )
-            elif command == "recipe-cat-full":
-                import tempfile
-
-                with tempfile.TemporaryDirectory() as d:
-                    session.run(
-                        "grayskull",
-                        "pypi",
-                        sdist_path,
-                        "-o",
-                        d,
-                    )
-                    session.run(
-                        "cat", str(Path(d) / PACKAGE_NAME / "meta.yaml"), external=True
-                    )
-
-            elif command == "build":
-                session.run(
-                    "conda",
-                    "mambabuild",
-                    "--output-folder=dist-conda/build",
-                    "--no-anaconda-upload",
-                    "dist-conda",
-                )
+nox.session(name="docs", **DEFAULT_KWS)(docs)
+nox.session(name="docs-conda", **CONDA_DEFAULT_KWS)(docs)
 
 
-def _append_recipe(recipe_path: str, append_path: str) -> None:
-    with open(recipe_path) as f:
-        recipe = f.readlines()
-
-    with open(append_path) as f:
-        append = f.readlines()
-
-    with open(recipe_path, "w") as f:
-        f.writelines(recipe + ["\n"] + append)
-
-
-# type checking
-def _typing(
+# ** lint
+@nox.session(python=False)
+def lint(
     session: nox.Session,
-    run: list[list[str]],
-    cmd: list[str],
-    run_internal: list[list[str]],
 ) -> None:
-    session_run_commands(session, run)
-    if not run and not run_internal and not cmd:
+    """
+    Run linters with pre-commit.
+
+    Defaults to `pre-commit run --all-files`.
+    To run something else pass, e.g.,
+    `nox -s lint -- --lint-run "pre-commit run --hook-stage manual --all-files`
+    """
+    pipxrun.run(
+        "pre-commit",
+        "run",
+        "--all-files",  # "--show-diff-on-failure",
+        specs=get_pipxrun_specs(),
+        session=session,
+    )
+
+
+# ** type checking
+@add_opts
+def typing(
+    session: nox.Session,
+    opts: SessionParams,
+) -> None:
+    """Run type checkers (mypy, pyright, pytype)."""
+    runner = (
+        Installer.from_envname(
+            session=session,
+            envname="typing",
+            lock=opts.lock,
+            update=opts.update,
+            # need package for nbqa checks
+            package=True,
+        )
+        .install_all(log_session=opts.log_session)
+        .run_commands(opts.typing_run)
+    )
+
+    cmd = opts.typing or []
+    if not opts.typing_run and not opts.typing_run_internal and not cmd:
         cmd = ["mypy", "pyright"]
 
     if "all" in cmd:
@@ -979,241 +932,295 @@ def _typing(
     # set the cache directory for mypy
     session.env["MYPY_CACHE_DIR"] = str(Path(session.create_tmp()) / ".mypy_cache")
 
-    def _run_info(cmd: str) -> None:
-        session.run("which", cmd, external=True)
-        session.run(cmd, "--version", external=True)
+    if "clean" in cmd:
+        cmd.remove("clean")
+
+        for name in [".mypy_cache", ".pytype"]:
+            p = Path(session.create_tmp()) / name
+            if p.exists():
+                session.log(f"removing cache {p}")
+                shutil.rmtree(str(p))
+
+    if not isinstance(session.python, str):
+        raise TypeError
+
+    run = partial(
+        pipxrun.run,
+        specs=get_pipxrun_specs(PIPXRUN_LOCK_REQUIREMENTS),
+        session=session,
+        python_version=session.python,
+        python_executable=runner.python_full_path,
+        external=True,
+    )
 
     for c in cmd:
-        if not c.startswith("nbqa"):
-            _run_info(c)
-        if c == "mypy":
-            session.run("mypy", "--color-output")
-        elif c == "pyright":
-            session.run("pyright", external=True)
-        elif c == "pytype":
-            session.run("pytype", "-o", str(Path(session.create_tmp()) / ".pytype"))
-        elif c.startswith("nbqa"):
+        if c.endswith("-notebook"):
             session.run("make", c, external=True)
+        elif c == "mypy":
+            run("mypy", "--color-output")
+        elif c == "pyright":
+            run("pyright")
         else:
-            session.log(f"skipping unknown command {c}")
-    session_run_commands(session, run_internal, external=False)
+            session.log(f"Skipping unknown command {c}")
+
+    for cmds in combine_list_list_str(opts.typing_run_internal or []):
+        run(*cmds)
+
+    # runner.run_commands(opts.typing_run_internal, external=False)
 
 
-@ALL_SESSION
-def typing(
-    session: nox.Session,
-    typing_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "mypy",
-            "pyright",
-            "pytype",
-            "all",
-            "nbqa-mypy",
-            "nbqa-pyright",
-            "nbqa-typing",
-        ],
-        flags=("--typing-cmd", "-m"),
-    ) = (),
-    typing_run: RUN_CLI = [],  # noqa
-    typing_run_internal: run_annotated(  # type: ignore
-        help="run arbitrary (internal) commands.  For example, --typing-run-internal 'mypy --some-option'",
-    ) = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
-) -> None:
-    """Run type checkers (mypy, pyright, pytype)."""
+nox.session(name="typing", **ALL_KWS)(typing)
+nox.session(name="typing-conda", **CONDA_ALL_KWS)(typing)
 
-    pkg_install_condaenv(
+
+# # ** Dist pypi
+@nox.session
+@add_opts
+def build(session: nox.Session, opts: SessionParams) -> None:
+    """
+    Build the distribution.
+
+    Note that default is to not use build isolation.
+    Pass `--build-isolation` to use build isolation.
+    """
+    runner = Installer.from_envname(
         session=session,
-        name="typing",
-        lock=lock,
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-    )
+        envname="build",
+        update=opts.update,
+    ).install_all(log_session=opts.log_session)
 
-    _typing(
-        session=session,
-        run=typing_run,
-        cmd=typing_cmd,
-        run_internal=typing_run_internal,
-    )
+    if opts.version:
+        session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = opts.version
 
+    for cmd in opts.build or ["build"]:
+        if cmd == "version":
+            session.run(runner.python_full_path, "-m", "hatchling", "version")
 
-@ALL_SESSION_VENV
-def typing_venv(
-    session: nox.Session,
-    typing_cmd: cmd_annotated(  # type: ignore
-        choices=[
-            "mypy",
-            "pyright",
-            "pytype",
-            "all",
-            "nbqa-mypy",
-            "nbqa-pyright",
-            "nbqa-typing",
-        ],
-        flags=("--typing-cmd", "-m"),
-    ) = (),
-    typing_run: RUN_CLI = [],  # noqa
-    typing_run_internal: run_annotated(  # type: ignore
-        help="run arbitrary (internal) commands.  For example, --typing-run-internal 'mypy --some-option'",
-    ) = [],  # noqa
-    lock: LOCK_CLI = False,
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    log_session: bool = False,
-) -> None:
-    """Run type checkers (mypy, pyright, pytype)."""
+        elif cmd == "build":
+            if Path(outdir := opts.build_outdir).exists():
+                shutil.rmtree(outdir)
 
-    pkg_install_venv(
-        session=session,
-        name="typing",
-        lock=lock,
-        install_package=True,
-        force_reinstall=force_reinstall,
-        log_session=log_session,
-        extras="typing",
-    )
+            args = f"{runner.python_full_path} -m build --outdir {outdir}".split()
+            if not opts.build_isolation:
+                args.append("--no-isolation")
 
-    _typing(
-        session=session,
-        run=typing_run,
-        cmd=typing_cmd,
-        run_internal=typing_run_internal,
-    )
+            if opts.build_opts:
+                args.extend(opts.build_opts)
+
+            out = session.run(*args, silent=opts.build_silent)
+            if opts.build_silent:
+                if not isinstance(out, str):
+                    msg = "session.run output not a string"
+                    raise ValueError(msg)
+                session.log(out.strip().split("\n")[-1])
 
 
-# ** testdist conda
-@ALL_SESSION
-def testdist_conda(
+def get_package_wheel(
     session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_conda_run: RUN_CLI = [],  # noqa
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Test conda distribution."""
+    opts: str | Iterable[str] | None = None,
+    extras: str | Iterable[str] | None = None,
+    reuse: bool = True,
+) -> str:
+    """
+    Build the package in return the build location
 
-    install_str = PACKAGE_NAME
-    if version:
-        install_str = f"{install_str}=={version}"
+    This is similar to how tox does isolated builds.
 
-    assert isinstance(session.python, str)
-    session_install_envs(
-        session,
-        session_environment_filename(
-            python_version=session.python, name="test-extras.yaml"
-        ),
-        deps=[install_str],
-        channels=["conda-forge"],
-        force_reinstall=force_reinstall,
-        install_package=False,
-    )
+    Note that the first time this is called,
 
-    if log_session:
-        session_log_session(session, False)
+    Should be straightforward to extend this to isolated builds
+    that depend on python version (something like have session build-3.11 ....)
+    """
+    dist_location = Path(session.cache_dir) / "dist"
+    if reuse and getattr(get_package_wheel, "_called", False):
+        session.log("Reuse isolated build")
+    else:
+        cmd = f"nox -s build -- ++build-outdir {dist_location} ++build-opts -w ++build-silent"
+        session.run_always(*shlex.split(cmd), external=True)
 
-    session_run_commands(session, testdist_conda_run)
+        # save that this was called:
+        if reuse:
+            get_package_wheel._called = True  # type: ignore[attr-defined]
 
-    if not test_no_pytest:
-        opts = combine_list_str(test_opts)
-        session.run("pytest", *opts)
+    paths = list(dist_location.glob("*.whl"))
+    if len(paths) != 1:
+        msg = f"something wonky with paths {paths}"
+        raise ValueError(msg)
 
+    path = str(paths[0])
 
-# ** testdist pypi
-@ALL_SESSION_VENV
-def testdist_pypi(
-    session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_pypi_run: RUN_CLI = [],  # noqa
-    testdist_pypi_extras: cmd_annotated(help="extras to install") = (),  # type: ignore
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
-) -> None:
-    """Test pypi distribution."""
-    extras = testdist_pypi_extras
-    install_str = PACKAGE_NAME
+    if cached_which("uv"):
+        path = f"{PACKAGE_NAME}@{path}"
 
     if extras:
-        install_str = "{}[{}]".format(install_str, ",".join(extras))
+        if not isinstance(extras, str):
+            extras = ",".join(extras)
+        path = f"{path}[{extras}]"
 
-    if version:
-        install_str = f"{install_str}=={version}"
+    if opts:
+        if not isinstance(opts, str):
+            opts = " ".join(opts)
+        path = f"{path} {opts}"
 
-    pkg_install_venv(
-        session=session,
-        name="testdist-pypi",
-        requirement_paths=[session_environment_filename(name="test-extras.txt")],
-        reqs=[install_str],
-        force_reinstall=force_reinstall,
-        install_package=False,
+    return path
+
+
+@nox.session(python=False)
+@add_opts
+def publish(session: nox.Session, opts: SessionParams) -> None:
+    """Publish the distribution"""
+    run = partial(
+        pipxrun.run, specs=get_pipxrun_specs(), session=session, external=True
     )
 
-    if log_session:
-        session_log_session(session, False)
-
-    _test(
-        session=session,
-        run=testdist_pypi_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=True,
-    )
+    for cmd in opts.publish or []:
+        if cmd == "test":
+            run("twine", "upload", "--repository", "testpypi", "dist/*")
+        elif cmd == "release":
+            run("twine", "upload", "dist/*")
+        elif cmd == "check":
+            run("twine", "check", "--strict", "dist/*")
 
 
-@ALL_SESSION
-def testdist_pypi_condaenv(
-    session: Session,
-    test_no_pytest: bool = False,
-    test_opts: TEST_OPTS_CLI = (),  # type: ignore
-    testdist_pypi_run: RUN_CLI = [],  # noqa
-    testdist_pypi_extras: cmd_annotated(help="extras to install") = (),  # type: ignore
-    force_reinstall: FORCE_REINSTALL_CLI = False,
-    version: VERSION_CLI = "",
-    log_session: bool = False,
+# # ** Dist conda
+@nox.session(name="conda-recipe", python=False)
+@add_opts
+def conda_recipe(
+    session: nox.Session,
+    opts: SessionParams,
 ) -> None:
-    """Test pypi distribution."""
-    extras = testdist_pypi_extras
-    install_str = PACKAGE_NAME
+    """Run grayskull to create recipe"""
+    commands = opts.conda_recipe or ["recipe"]
 
-    if extras:
-        install_str = "{}[{}]".format(install_str, ",".join(extras))
+    run = partial(pipxrun.run, specs=get_pipxrun_specs(), session=session)
 
-    if version:
-        install_str = f"{install_str}=={version}"
+    if not (sdist_path := opts.conda_recipe_sdist_path):
+        sdist_path = PACKAGE_NAME
+        if opts.version:
+            sdist_path = f"{sdist_path}=={opts.version}"
 
-    assert isinstance(session.python, str)
-    session_install_envs(
-        session,
-        session_environment_filename(
-            python_version=session.python, name="test-extras.yaml"
-        ),
-        reqs=[install_str],
-        channels=["conda-forge"],
-        force_reinstall=force_reinstall,
-        install_package=False,
-    )
-    if log_session:
-        session_log_session(session, False)
+    for command in commands:
+        if command == "recipe":
+            # make directory?
+            if not (d := Path("./dist-conda")).exists():
+                d.mkdir()
 
-    _test(
+            run(
+                "grayskull",
+                "pypi",
+                sdist_path,
+                "--sections",
+                "package",
+                "source",
+                "build",
+                "requirements",
+                "-o",
+                "dist-conda",
+            )
+
+            _append_recipe(
+                f"dist-conda/{PACKAGE_NAME}/meta.yaml",
+                "config/recipe-append.yaml",
+            )
+            session.run("cat", f"dist-conda/{PACKAGE_NAME}/meta.yaml", external=True)
+
+        elif command == "recipe-full":
+            import tempfile
+
+            with tempfile.TemporaryDirectory() as d:  # type: ignore[assignment,unused-ignore]
+                run(
+                    "grayskull",
+                    "pypi",
+                    sdist_path,
+                    "-o",
+                    str(d),
+                )
+                # session.run(
+                #     sys.executable,
+                #     "tools/pipxrun.py",
+                #     PIPXRUN_REQUIREMENTS,
+                #     "-v",
+                #     "-c",
+                #     " ".join(
+                #         [
+                #             "grayskull",
+                #             "pypi",
+                #             sdist_path,
+                #             "-o",
+                #             str(d),
+                #         ]
+                #     ),
+                # )
+                path = Path(d) / PACKAGE_NAME / "meta.yaml"
+                session.log(f"cat {path}:")
+                with path.open() as f:
+                    for line in f:
+                        print(line, end="")  # noqa: T201
+
+                # # session.run(
+                # #     "cat",
+                #     str(Path(d) / PACKAGE_NAME / "meta.yaml"),
+                #     external=True,
+                # )
+
+
+@nox.session(name="conda-build", **CONDA_DEFAULT_KWS)
+@add_opts
+def conda_build(session: nox.Session, opts: SessionParams) -> None:
+    """Run `conda mambabuild`."""
+    runner = Installer.from_envname(
         session=session,
-        run=testdist_pypi_run,
-        test_no_pytest=test_no_pytest,
-        test_opts=test_opts,
-        no_cov=True,
-    )
+        update=opts.update,
+        conda_deps=["boa", "anaconda-client"],
+    ).install_all(log_session=opts.log_session)
+
+    cmds, run = opts.conda_build, opts.conda_build_run
+
+    runner.run_commands(run)
+
+    if not run and not cmds:
+        cmds = ["build", "clean"]
+
+    if cmds is None:
+        cmds = []
+
+    if "clean" in cmds:
+        cmds.remove("clean")
+        session.log("removing directory dist-conda/build")
+        if Path("./dist-conda/build").exists():
+            shutil.rmtree("./dist-conda/build")
+
+    for cmd in cmds:
+        if cmd == "build":
+            if not (d := Path(f"./dist-conda/{PACKAGE_NAME}/meta.yaml")).exists():
+                msg = f"no file {d}"
+                raise ValueError(msg)
+
+            session.run(
+                "conda",
+                "mambabuild",
+                "--output-folder=dist-conda/build",
+                "--no-anaconda-upload",
+                "dist-conda",
+            )
 
 
-# * Utilities --------------------------------------------------------------------------
+# ** Other utilities
+@nox.session
+@add_opts
+def cog(session: nox.Session, opts: SessionParams) -> None:
+    """Run cog."""
+    Installer.from_envname(
+        session=session,
+        update=opts.update,
+        pip_deps="cogapp",
+    ).install_all(log_session=opts.log_session)
+    session.run("cog", "-rP", "README.md", env={"COLUMNS": "90"})
+
+
+# * Utilities -------------------------------------------------------------------------
 def _create_doc_examples_symlinks(session: nox.Session, clean: bool = True) -> None:
     """Create symlinks from docs/examples/*.md files to /examples/usage/..."""
-
     import os
 
     def usage_paths(path: Path) -> Iterator[Path]:
@@ -1229,17 +1236,20 @@ def _create_doc_examples_symlinks(session: nox.Session, clean: bool = True) -> N
     ) -> Path:
         path = Path(prefix_dir) / Path(usage_path)
 
-        assert all(ext.startswith(".") for ext in exts)
+        if not all(ext.startswith(".") for ext in exts):
+            msg = "Bad extensions.  Should start with '.'"
+            raise ValueError(msg)
 
         if path.exists():
             return path
-        else:
-            for ext in exts:
-                p = path.with_suffix(ext)
-                if p.exists():
-                    return p
 
-        raise ValueError(f"no path found for base {path}")
+        for ext in exts:
+            p = path.with_suffix(ext)
+            if p.exists():
+                return p
+
+        msg = f"no path found for base {path}"
+        raise ValueError(msg)
 
     root = Path("./docs/examples/")
     if clean:
@@ -1267,33 +1277,15 @@ def _create_doc_examples_symlinks(session: nox.Session, clean: bool = True) -> N
             os.symlink(target_rel, link)
 
 
-# # If want separate env for updating/reporting version with setuptools-scm
-# # We do this from dev environment.
-# # ** version report/update
-# @DEFAULT_SESSION_VENV
-# def version_scm(
-#     session: Session,
-#     version: VERSION_CLI = "",
-#     force_reinstall: FORCE_REINSTALL_CLI = False,
-# ):
-#     """
-#     Get current version from setuptools-scm
+def _append_recipe(recipe_path: str | Path, append_path: str | Path) -> None:
+    recipe_path = Path(recipe_path)
+    append_path = Path(append_path)
 
-#     Note that the version of editable installs can get stale.
-#     This will show the actual current version.
-#     Avoids need to include setuptools-scm in develop/docs/etc.
-#     """
+    with recipe_path.open() as f:
+        recipe = f.readlines()
 
-#     pkg_install_venv(
-#         session=session,
-#         name="version-scm",
-#         install_package=True,
-#         reqs=["setuptools_scm"],
-#         force_reinstall=force_reinstall,
-#         no_deps=True,
-#     )
+    with append_path.open() as f:
+        append = f.readlines()
 
-#     if version:
-#         session.env["SETUPTOOLS_SCM_PRETEND_VERSION"] = version
-
-#     session.run("python", "-m", "setuptools_scm")
+    with recipe_path.open("w") as f:
+        f.writelines([*recipe, "\n", *append])
