@@ -1,6 +1,6 @@
 $(document).ready(function () {
     // Mark external (non-nist.gov) A tags with class "external"
-    //If the adress start with https and ends with nist.gov
+    //If the address start with https and ends with nist.gov
     var re_nist = new RegExp("^https?://((^/)*.)*nist\\.gov(/|$)");
     //Regex to find address that start with https
     var re_absolute_address = new RegExp("^((https?:)?//)");
