@@ -1,2 +1,3 @@
 """Interface to :mod:`lnpy.stability`"""
-from lnpy.stability import *  # noqa: F403, F401 # pyright: ignore
+
+from lnpy.stability import *  # noqa: F403 # pyright: ignore
