@@ -8,13 +8,14 @@ from __future__ import annotations
 try:
     import importlib_resources as resources
 except ImportError:
-    import importlib.resources as resources  # type: ignore[no-redef]
+    from importlib import resources  # type: ignore[no-redef]
 
 import json
 from dataclasses import asdict, dataclass
 from typing import TYPE_CHECKING, TypedDict
 
-from ._lazy_imports import np, xr
+import numpy as np
+import xarray as xr
 
 # from .lnpiseries import lnPiCollection
 from .segment import BuildPhasesBase, PhaseCreator
@@ -53,10 +54,8 @@ def json_to_dict(basename: str) -> dict[str, Any]:
     else:
         fopen = open  # type: ignore[assignment]
 
-    with fopen(resources.files("lnpy.data").joinpath(basename), "r") as f:
-        out = json.load(f)
-
-    return out  # type: ignore
+    with fopen(resources.files("lnpy.data").joinpath(basename), "r") as f:  # pyright: ignore[reportCallIssue,reportArgumentType]
+        return json.load(f)  # type: ignore[no-any-return]
 
 
 class ExampleDict(TypedDict):
@@ -105,15 +104,16 @@ def load_example_lnpimasked(name: _ExampleNames) -> lnPiMasked:
         "watermof": "json",
     }
 
-    assert name in extensions
+    if name not in extensions:
+        msg = f"{name=} not in {extensions=}"
+        raise ValueError(msg)
 
     basename = f"{name}_example.{extensions[name]}"
 
     d = json_to_dict(basename)
     ds = xr.Dataset.from_dict(d)
 
-    ref = dataset_to_lnpimasked(ds)
-    return ref
+    return dataset_to_lnpimasked(ds)
 
 
 @dataclass
@@ -157,7 +157,8 @@ def lj_sup_example() -> Example:
 
 def tag_phases_single_comp_simple(x: Sequence[lnPiMasked]) -> MyNDArray:
     if len(x) > 2:
-        raise ValueError("bad tag function")
+        msg = "bad tag function"
+        raise ValueError(msg)
     argmax0 = np.array([xx.local_argmax()[0] for xx in x])
     return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
 
@@ -197,7 +198,8 @@ def hsmix_example() -> Example:
 
     def tag_phases(x: Sequence[lnPiMasked]) -> MyNDArray:
         if len(x) > 2:
-            raise ValueError("bad tag function")
+            msg = "bad tag function"
+            raise ValueError(msg)
         argmax0 = np.array([xx.local_argmax()[0] for xx in x])
         return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
 
