@@ -2,11 +2,12 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
 
-from ._lazy_imports import np
+import numpy as np
+
 from .lnpiseries import lnPiCollection
-from .utils import RootResultDict, rootresults_to_rootresultdict
+from .utils import RootResultDict, array_to_scalar, rootresults_to_rootresultdict
 
 if TYPE_CHECKING:
     from typing import Any, Mapping, Sequence
@@ -40,18 +41,19 @@ def tag_phases_singlecomp(x: Sequence[lnPiMasked]) -> Sequence[int] | MyNDArray:
         Phase code for each element in ``x``.
     """
     if len(x) > 2:
-        raise ValueError("bad tag function")
-    else:
-        argmax0 = np.array([xx.local_argmax()[0] for xx in x])
-        if len(x) == 2:
-            return np.argsort(argmax0)
-        else:
-            return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
+        msg = "bad tag function"
+        raise ValueError(msg)
+
+    argmax0 = np.array([xx.local_argmax()[0] for xx in x])
+    if len(x) == 2:
+        return np.argsort(argmax0)
+
+    return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
 
 
 def get_lnz_min(
     target: float,
-    C: lnPiCollection,
+    collection: lnPiCollection,
     build_phases: BuildPhasesBase,
     phase_id: int | None = 0,
     component: int | None = None,
@@ -69,7 +71,7 @@ def get_lnz_min(
     ----------
     target : float
         target density of specified component
-    C : CollectionPhases
+    collection : CollectionPhases
         initial guess to work from.  This is assumed to be sorted in ascending `lnz` order
     build_phases : callable
         function to build new phases objects
@@ -80,7 +82,7 @@ def get_lnz_min(
         If component is None, use build_phases.index.
         If component=='None', then consider total density
     dlnz : float, default=0.5
-        how to change the chemical potential if `C` doesn't already
+        how to change the chemical potential if `collection` doesn't already
         bracket solution
     ref : lnPiMasked, optional
         optional lnPiMasked object to pass to build_phases
@@ -99,8 +101,11 @@ def get_lnz_min(
     """
     from scipy.optimize import brentq
 
-    if phase_id is not None and phase_id not in C.index.get_level_values("phase"):
-        raise ValueError(f"no phase {phase_id}")
+    if phase_id is not None and phase_id not in collection.index.get_level_values(
+        "phase"
+    ):
+        msg = f"no phase {phase_id}"
+        raise ValueError(msg)
 
     lnz_idx = build_phases.index
     if isinstance(component, str) and component.lower() == "none":
@@ -124,7 +129,7 @@ def get_lnz_min(
         def getter(x: lnPiCollection) -> xr.DataArray:
             return getter_comp(x).sel(phase=phase_id)
 
-    s = getter(C).to_series().dropna()
+    s = getter(collection).to_series().dropna()
     # builder
     if build_kws is None:
         build_kws = {}
@@ -133,31 +138,34 @@ def get_lnz_min(
     left = None
     ss = s[s < target]
     if len(ss) > 0:
-        left = C.mloc[ss.index[[-1]]]
+        left = collection.mloc[ss.index[[-1]]]
     else:
-        new_lnz = C.mloc[s.index[[0]]]._get_lnz(lnz_idx)
+        # TODO(wpk): Need to make collection work better.  Might need to just have a single class for everything...
+        # reveal_type(collection.mloc[s.index[[0]]])
+        new_lnz = collection.mloc[s.index[[0]]]._get_lnz(lnz_idx)
         dlnz_left = dlnz
-        for i in range(ntry):
+        for _i in range(ntry):
             new_lnz -= dlnz_left
             p = build_phases(new_lnz, ref=ref, **build_kws)
-            if phase_id in p._get_level("phase") and getter(p).values < target:
+            if phase_id in p._get_level("phase") and getter(p).to_numpy() < target:
                 left = p
                 break
             dlnz_left *= dfac
 
     if left is None:
-        raise RuntimeError("could not find left bounds")
+        msg = "could not find left bounds"
+        raise RuntimeError(msg)
 
     # right bracket
     right = None
     ss = s[s > target]
     if len(ss) > 0:
-        right = C.mloc[ss.index[[0]]]
+        right = collection.mloc[ss.index[[0]]]
     else:
-        new_lnz = C.mloc[s.index[[-1]]]._get_lnz(lnz_idx)
+        new_lnz = collection.mloc[s.index[[-1]]]._get_lnz(lnz_idx)
         dlnz_right = dlnz
 
-        for i in range(ntry):
+        for _i in range(ntry):
             new_lnz += dlnz_right
             p = build_phases(new_lnz, ref=ref, **build_kws)
 
@@ -166,20 +174,21 @@ def get_lnz_min(
                 new_lnz -= dlnz_right
                 # reset to half dlnz
                 dlnz_right = dlnz_right * 0.5
-            elif getter(p).values > target:
+            elif getter(p).to_numpy() > target:
                 right = p
                 break
             else:
                 dlnz_right *= dfac
 
     if right is None:
-        raise RuntimeError("could not find right bounds")
+        msg = "could not find right bounds"
+        raise RuntimeError(msg)
 
-    def f(x: float) -> MyNDArray:
+    def f(x: float) -> float:
         lnz_new = x
         p = build_phases(lnz_new, ref=ref, **build_kws)
-        f.lnpi = p  # type: ignore
-        return getter(p).values - target
+        f.lnpi = p  # type: ignore[attr-defined]
+        return array_to_scalar(getter(p).values) - target
 
     a, b = sorted([x._get_lnz(lnz_idx) for x in [left, right]])
 
@@ -193,7 +202,7 @@ def get_lnz_max(
     build_phases: BuildPhasesBase,
     ref: lnPiMasked | None = None,
     lnz_start: float | None = None,
-    C: lnPiCollection | None = None,
+    collection: lnPiCollection | None = None,
     dlnz: float = 0.5,
     dfac: float = 1.0,
     threshold_abs: float = 1e-4,
@@ -206,45 +215,47 @@ def get_lnz_max(
     ref = ref or build_phases.phase_creator.ref
 
     if ref is None:
-        raise ValueError(
-            "must specify `ref` or build_phases must have access to reference lnPiMasked object"
-        )
+        msg = "must specify `ref` or build_phases must have access to reference lnPiMasked object"
+        raise ValueError(msg)
 
     lnz_idx = build_phases.index
-    lnz_start = lnz_start or ref.lnz[lnz_idx]
+    lnz_start = cast(float, lnz_start or ref.lnz[lnz_idx])
 
     # need left/right bounds
     # left is greatest lnz point with edge_distance > edge_distance_min
     # right is least lnz point with edge_distance < edge_distance_min
-    left = right = None
-    lnz_left = lnz_right = None
-    n_left = n_right = None
+    left: lnPiCollection | None = None
+    right: lnPiCollection | None = None
+    lnz_left: float | None = None
+    lnz_right: float | None = None
+    n_left: int | None = None
+    n_right: int | None = None
 
     def getter(p: lnPiCollection) -> xr.DataArray:
         v = p.xge.edge_distance(ref)
         if not p._xarray_unstack:
-            v = v.unstack(p._concat_dim)
+            v = v.unstack(p._concat_dim)  # noqa: PD010
         return v.min("phase")
 
-    # if have C, try to use it
-    if C is not None:
-        # see if bound contained in C
+    # if have collection, try to use it
+    if collection is not None:
+        # see if bound contained in collection
 
-        s = getter(C).to_series()
+        s = getter(collection).to_series()
 
         # left
         ss = s[s > edge_distance_min]
         if len(ss) > 0:
-            left = C.mloc[ss.index[[-1]]]
+            left = collection.mloc[ss.index[[-1]]]
         else:
-            lnz_left = C.zloc[[0]]._get_lnz(lnz_idx)
+            lnz_left = collection.zloc[[0]]._get_lnz(lnz_idx)
 
         # right
         ss = s[s < edge_distance_min]
         if len(ss) > 0:
-            right = C.mloc[ss.index[[0]]]
+            right = collection.mloc[ss.index[[0]]]
         else:
-            lnz_right = C.zloc[[-1]]._get_lnz(lnz_idx)
+            lnz_right = collection.zloc[[-1]]._get_lnz(lnz_idx)
 
     # if left not set, try to find it
     if left is None:
@@ -254,7 +265,7 @@ def get_lnz_max(
         for i in range(ntry):
             lnz_left -= dlnz_loc
             p = build_phases(lnz_left, ref=ref, **build_kws)
-            if getter(p).values >= edge_distance_min:
+            if getter(p).to_numpy() >= edge_distance_min:
                 left = p
                 n_left = i
                 break
@@ -262,7 +273,8 @@ def get_lnz_max(
             dlnz_loc *= dfac
 
     if left is None:
-        raise RuntimeError("could not find left bound")
+        msg = "could not find left bound"
+        raise RuntimeError(msg)
 
     # if right not set, try to find it
     if right is None:
@@ -274,7 +286,7 @@ def get_lnz_max(
             lnz_right += dlnz_loc
             p = build_phases(lnz_right, ref=ref, **build_kws)
 
-            if getter(p).values < edge_distance_min:
+            if getter(p).to_numpy() < edge_distance_min:
                 right = p
                 n_right = i
                 break
@@ -282,227 +294,31 @@ def get_lnz_max(
             dlnz_loc *= dfac
 
     if right is None:
-        raise RuntimeError("could not find right bound")
+        msg = "could not find right bound"
+        raise RuntimeError(msg)
 
     # not do bisection
-    P = [left, right]
-    Y = [getter(x).values for x in P]
+    bracket = [left, right]
+    values = [getter(x).to_numpy() for x in bracket]
 
-    for i in range(ntry):
-        lnz = [x._get_lnz(lnz_idx) for x in P]
+    for i in range(ntry):  # noqa: B007
+        lnz = [x._get_lnz(lnz_idx) for x in bracket]
         delta = np.abs(lnz[1] - lnz[0])
         if delta < threshold_abs:
             break
         lnz_mid = 0.5 * (lnz[0] + lnz[1])
         mid = build_phases(lnz_mid, ref=ref, **build_kws)
-        y_mid = getter(mid).values
+        y_mid = getter(mid).to_numpy()
 
-        if y_mid >= edge_distance_min:
-            index = 0
-        else:
-            index = 1
+        index = 0 if y_mid >= edge_distance_min else 1
 
-        P[index] = mid
-        Y[index] = y_mid
+        bracket[index] = mid
+        values[index] = y_mid
 
-    left, right = P
+    left, right = bracket
 
-    info = dict(ntry=i, ntry_left=n_left, ntry_right=n_right, precision=delta)
+    info = {"ntry": i, "ntry_left": n_left, "ntry_right": n_right, "precision": delta}
     return left, info
-
-
-# class _BaseLimit:
-#     _bound_side = None  # -1 for lower bound, +1 for upper bound
-
-#     def __init__(self, parent):
-#         self._parent = parent
-
-#     @property
-#     def parent(self):
-#         return self._parent
-
-#     def __call__(self, *args, **kwargs):
-#         raise NotImplementedError("to be implemented in subclass")
-
-#     @property
-#     def lnz_idx(self):
-#         return self._lnz_idx
-
-#     @lnz_idx.setter
-#     def lnz_idx(self, i):
-#         self._lnz_idx = i
-
-#     @property
-#     def access(self):
-#         return self._access
-
-#     @access.setter
-#     def access(self, val):
-#         if not isinstance(val, lnPiCollection):
-#             raise ValueError("access should be lnPiCollection")
-#         self._access = val
-
-#     def _get_bound(self, sign, access=None, lnz_idx=None):
-#         if access is None:
-#             access = self.access
-#         if lnz_idx is None:
-#             lnz_idx = self.lnz_idx
-#         if sign == self._bound_side:
-#             return access._get_lnz(lnz_idx)
-#         else:
-#             return sign * np.inf
-
-#     def get_bounds(self, access=None, lnz_idx=None):
-#         return np.array(
-#             [
-#                 self._get_bound(sign=sign, access=access, lnz_idx=lnz_idx)
-#                 for sign in [-1, +1]
-#             ]
-#         )
-
-#     @property
-#     def bounds(self):
-#         return self._bounds
-
-#     @bounds.setter
-#     def bounds(self, bounds):
-#         assert len(bounds) == 2
-#         self._bounds = np.array(bounds)
-
-#     def set_bounds(self, access=None, lnz_idx=None):
-#         self.bounds = self.get_bounds(access=access, lnz_idx=lnz_idx)
-
-#     @property
-#     def lb(self):
-#         return self.bounds[0]
-
-#     @property
-#     def ub(self):
-#         return self.bounds[1]
-
-#     @property
-#     def lnz_list(self):
-#         return self._lnz_list
-
-#     def get_lnz_list(self, lnz=None, lnz_idx=None):
-#         if lnz is None:
-#             lnz = self.access.iloc[0].lnz
-#         if lnz_idx is None:
-#             lnz_idx = self.lnz_idx
-#         lnz = list(lnz)
-#         lnz[lnz_idx] = None
-#         return lnz
-
-#     def set_lnz_list(self, lnz=None, lnz_idx=None):
-#         self._lnz_list = self.get_lnz_list(lnz, lnz_idx)
-
-
-# @lnPiCollection.decorate_accessor("bounds_lower", single_create=True)
-# class LowerBounds(_BaseLimit):
-#     """Class to hold lower bounds."""
-
-#     _bound_side = -1
-
-#     def __call__(
-#         self,
-#         dens_min,
-#         build_phases,
-#         phase_id=0,
-#         component=None,
-#         dlnz=0.5,
-#         dfac=1.0,
-#         ref=None,
-#         build_kws=None,
-#         ntry=20,
-#         solve_kws=None,
-#         inplace=True,
-#         force=False,
-#     ):
-#         if hasattr(self, "_access") and not force:
-#             p = self.access
-#             info = self._info
-#             lnz_idx = self.lnz_idx
-#             bounds = self.bounds
-
-#         else:
-#             lnz_idx = build_phases.index
-#             p, info = get_lnz_min(
-#                 target=dens_min,
-#                 C=self._parent,
-#                 build_phases=build_phases,
-#                 phase_id=phase_id,
-#                 component=component,
-#                 dlnz=dlnz,
-#                 dfac=1.0,
-#                 ref=ref,
-#                 build_kws=build_kws,
-#                 ntry=ntry,
-#                 solve_kws=solve_kws,
-#             )
-
-#             bounds = self.get_bounds(p, lnz_idx)
-
-#         if inplace:
-#             self.access = p
-#             self._info = info
-#             self.lnz_idx = lnz_idx
-#             self.bounds = bounds
-#             self.set_lnz_list()
-#         else:
-#             return bounds, p, lnz_idx, info
-
-
-# @lnPiCollection.decorate_accessor("bounds_upper", single_create=True)
-# class UpperBounds(_BaseLimit):
-#     """Class to hold upper bounds."""
-
-#     _bound_side = +1
-
-#     def __call__(
-#         self,
-#         edge_distance_min,
-#         build_phases,
-#         ref=None,
-#         lnz_start=None,
-#         dlnz=0.5,
-#         dfac=1.0,
-#         threshold_abs=1e-4,
-#         ntry=20,
-#         build_kws=None,
-#         inplace=True,
-#         force=False,
-#     ):
-#         if hasattr(self, "_access") and not force:
-#             p = self.access
-#             info = self._info
-#             lnz_idx = self.lnz_idx
-#             bounds = self.bounds
-
-#         else:
-#             lnz_idx = build_phases.index
-#             p, info = get_lnz_max(
-#                 edge_distance_min=edge_distance_min,
-#                 C=self._parent,
-#                 build_phases=build_phases,
-#                 lnz_start=None,
-#                 dlnz=dlnz,
-#                 dfac=1.0,
-#                 threshold_abs=threshold_abs,
-#                 ntry=ntry,
-#                 ref=ref,
-#                 build_kws=build_kws,
-#                 full_output=True,
-#             )
-#             bounds = self.get_bounds(p, lnz_idx)
-
-#         if inplace:
-#             self.access = p
-#             self._info = info
-#             self.lnz_idx = lnz_idx
-#             self.bounds = bounds
-#             self.set_lnz_list()
-#         else:
-#             return bounds, p, lnz_idx, info
 
 
 def build_grid(
@@ -546,18 +362,22 @@ def build_grid(
     """
 
     if x is None:
-        if (dx is None) or (x_range is None and offsets is None):
-            raise ValueError(
-                "must specify dx and one of x_range or offsets if not passing x"
-            )
+        if dx is None:
+            msg = "Must specify dx"
+            raise ValueError(msg)
 
-        if x_range is None:
-            x_range = x0 + np.array(offsets)
+        if x_range is not None:
+            x_range = np.asarray(x_range)
+        elif offsets is not None and x0 is not None:
+            x_range = x0 + np.asarray(offsets)
         else:
-            x_range = np.asarray(x)
-        assert len(x_range) == 2
+            msg = "must specify x_range or x0 and offsets"
+            raise ValueError(msg)
 
-        # for consistancy with previous code.
+        if len(x_range) != 2:
+            raise ValueError
+
+        # for consistency with previous code.
         if even_grid:
             new_range = np.round(x_range / dx) * dx
             if new_range[0] < x_range[0]:
@@ -565,7 +385,7 @@ def build_grid(
             if new_range[1] > x_range[1]:
                 new_range[1] -= dx
             x_range = new_range
-        x = np.arange(x_range[0], x_range[1] + dx * 0.5, dx)  # type: ignore
+        x = np.arange(x_range[0], x_range[1] + dx * 0.5, dx)  # type: ignore[index]
     else:
         x = np.asarray(x)
 
@@ -573,19 +393,20 @@ def build_grid(
         lb, ub = x.min(), x.max()
 
     if even_grid:
-        assert dx is not None
+        if dx is None:
+            raise TypeError
         x = np.round(x / dx) * dx
 
     if digits is not None:
-        x = np.round(x, digits)  # type: ignore
+        x = np.round(x, digits)  # type: ignore[arg-type]
 
     if unique:
-        x = np.unique(x)  # type: ignore
+        x = np.unique(x)  # type: ignore[arg-type]
 
     if not outlier:
-        x = x[(lb <= x) & (x <= ub)]  # type: ignore
+        x = x[(lb <= x) & (x <= ub)]  # type: ignore[index]
 
-    return x  # type: ignore
+    return x  # type: ignore[return-value]
 
 
 def limited_collection(
@@ -602,7 +423,7 @@ def limited_collection(
     lnz_max_kws: Mapping[str, Any] | None = None,
     ref: lnPiMasked | None = None,
     build_kws: Mapping[str, Any] | None = None,
-    build_stability_kws: Mapping[str, Any] | None = None,
+    # build_stability_kws: Mapping[str, Any] | None = None,
     nmax: int | None = None,
     xarray_output: bool = True,
     collection_kws: Mapping[str, Any] | None = None,
@@ -619,14 +440,13 @@ def limited_collection(
     if lnz_range is None:
         ref = ref or build_phases.phase_creator.ref
         if ref is None:
-            raise ValueError(
-                "Must pass in ref or build_phases must have access to reference lnPiMasked object"
-            )
+            msg = "Must pass in ref or build_phases must have access to reference lnPiMasked object"
+            raise ValueError(msg)
         x0 = ref.lnz[build_phases.index]
     else:
         x0 = None
 
-    # TODO: update tests to get rid of outlier=True
+    # TODO: update tests to get rid of outlier=True  # noqa: TD002
     lnzs = build_grid(
         dx=dlnz,
         x_range=lnz_range,
@@ -664,7 +484,7 @@ def limited_collection(
                 )
                 if o["converged"]:
                     lnz_min = p_min.iloc[0].lnz[build_phases.index] - dlnz
-            except Exception:
+            except Exception:  # noqa: BLE001, S110
                 pass
 
         if edge_distance_min is not None:
@@ -675,11 +495,11 @@ def limited_collection(
                     edge_distance_min,
                     build_phases,
                     build_kws=build_kws,
-                    C=c_course,
+                    collection=c_course,
                     **lnz_max_kws,
                 )
                 lnz_max = p_max.iloc[0].lnz[build_phases.index]
-            except Exception:
+            except Exception:  # noqa: BLE001,S110
                 pass
 
     lnzs = lnzs[(lnzs >= lnz_min) & (lnzs <= lnz_max)]
