@@ -1,5 +1,6 @@
 # mypy: disable-error-code="no-untyped-def, no-untyped-call"
 import json
+import locale
 from pathlib import Path
 
 import numpy as np
@@ -14,13 +15,17 @@ path_data = Path(__file__).parent / "water_cluster"
 
 def tag_phases2(x):
     if len(x) > 2:
-        raise ValueError("bad tag function")
+        msg = "bad tag function"
+        raise ValueError(msg)
     argmax0 = np.array([xx.local_argmax()[0] for xx in x])
     return np.where(argmax0 <= x[0].shape[0] / 2, 0, 1)
 
 
 def load_ref():
-    ensemble = json.load(open(path_data / "water_MOF_ensemble.json"))
+    with Path(path_data / "water_MOF_ensemble.json").open(
+        encoding=locale.getpreferredencoding(False)
+    ) as f:
+        ensemble = json.load(f)
     lnz = ensemble["betamu0"]
     box = ensemble["box"]
     state_kws = {
@@ -31,12 +36,12 @@ def load_ref():
     data = pd.read_csv(path_data / "water_MOF_example.csv")
 
     # potential energy
-    pe = data["U"].values
+    pe = data["U"].to_numpy()
 
     # Build the lnPi object
     # DWS Note to self: I'm doing something sloppy here, in that the N values are not specified
-    ref = lnpy.lnPiMasked.from_data(
-        data=data["lnPi"].values,  # type: ignore
+    return lnpy.lnPiMasked.from_data(
+        data=data["lnPi"].values,  # type: ignore[arg-type]
         fill_value=np.nan,
         lnz=lnz,
         lnz_data=lnz,
@@ -45,17 +50,16 @@ def load_ref():
         # extra_kws is where you pass things which will be passed along to new lnPis, like potential energy
         extra_kws={"PE": pe, "other": pe},
     )
-    return ref
 
 
 @pytest.fixture(params=[0, 1])
 def ref(request):
     if request.param == 0:
         return load_ref()
-    else:
-        import lnpy.examples
 
-        return lnpy.examples.load_example_lnpimasked("watermof")
+    import lnpy.examples
+
+    return lnpy.examples.load_example_lnpimasked("watermof")
 
 
 def get_test_table(o, ref):
@@ -78,7 +82,7 @@ def get_test_table(o, ref):
     )  # .to_csv('data_0.csv', index=False)
 
 
-def do_test(phase_creator, ref, fname):
+def do_test(phase_creator, ref, fname) -> None:
     p = phase_creator.build_phases()
 
     other = get_test_table(p, ref)
@@ -87,7 +91,7 @@ def do_test(phase_creator, ref, fname):
     pd.testing.assert_frame_equal(other, test)
 
 
-def test_0(ref):
+def test_0(ref) -> None:
     phase_creator = lnpy.segment.PhaseCreator(
         nmax=2,  # number of phases
         nmax_peak=10,  # max number of peaks in lnPi
@@ -98,7 +102,7 @@ def test_0(ref):
     do_test(phase_creator, ref, "data_0.csv")
 
 
-def test_1(ref):
+def test_1(ref) -> None:
     phase_creator = lnpy.segment.PhaseCreator(
         nmax=20,  # number of phases
         nmax_peak=50,  # max number of peaks in lnPi
